adds MKCoordinateRegion extension for fitting map to items
