--- PhotoPoints/View/Map/Map Feature Data/Streams.swift
@@ -1,619 +1,0 @@
-//
-//  Streams.swift
-//  CoordinateDrawer
-//
-//  Created by Clay Suttner on 12/22/20.
-//
-
-import Foundation
-import MapKit
-
-class StreamLine: MKPolyline {}
-
-let streams: [[CLLocationCoordinate2D]] = [
-[
-CLLocationCoordinate2D(latitude:47.7727657186765,longitude:-122.192386887836),
-CLLocationCoordinate2D(latitude:47.7727633390059,longitude:-122.192440010658),
-CLLocationCoordinate2D(latitude:47.7727627440882,longitude:-122.192492911603),
-CLLocationCoordinate2D(latitude:47.772745689782,longitude:-122.192538951381),
-CLLocationCoordinate2D(latitude:47.7727222896873,longitude:-122.192578129993),
-CLLocationCoordinate2D(latitude:47.7727185218755,longitude:-122.192631030938),
-CLLocationCoordinate2D(latitude:47.7727185218755,longitude:-122.192685261016),
-CLLocationCoordinate2D(latitude:47.7727068218281,longitude:-122.192737055767),
-CLLocationCoordinate2D(latitude:47.7726921471925,longitude:-122.192786416252),
-CLLocationCoordinate2D(latitude:47.7726750928863,longitude:-122.192834890296),
-CLLocationCoordinate2D(latitude:47.7726572453564,longitude:-122.192888898497),
-CLLocationCoordinate2D(latitude:47.7726386046031,longitude:-122.19293383102),
-CLLocationCoordinate2D(latitude:47.7726231367439,longitude:-122.192982306125),
-CLLocationCoordinate2D(latitude:47.7726116350025,longitude:-122.193032993619),
-CLLocationCoordinate2D(latitude:47.7726092553318,longitude:-122.193100726278),
-CLLocationCoordinate2D(latitude:47.7726076688847,longitude:-122.1931538491),
-CLLocationCoordinate2D(latitude:47.7726023146258,longitude:-122.193206750045),
-CLLocationCoordinate2D(latitude:47.7726015214022,longitude:-122.193285106206),
-CLLocationCoordinate2D(latitude:47.7725975552845,longitude:-122.193339114408),
-CLLocationCoordinate2D(latitude:47.7725929942491,longitude:-122.193392237229),
-CLLocationCoordinate2D(latitude:47.7725922010255,longitude:-122.193445139236),
-CLLocationCoordinate2D(latitude:47.7726068756612,longitude:-122.193503795154),
-CLLocationCoordinate2D(latitude:47.7726161960379,longitude:-122.193555589905),
-CLLocationCoordinate2D(latitude:47.7726215502968,longitude:-122.193608712726),
-CLLocationCoordinate2D(latitude:47.7726263096381,longitude:-122.193663828184),
-CLLocationCoordinate2D(latitude:47.7726308706735,longitude:-122.193716950474),
-CLLocationCoordinate2D(latitude:47.7726308706735,longitude:-122.193770958676),
-CLLocationCoordinate2D(latitude:47.7726300774499,longitude:-122.193823860152),
-CLLocationCoordinate2D(latitude:47.772624723191,longitude:-122.193876982974),
-CLLocationCoordinate2D(latitude:47.7726185757085,longitude:-122.193933204626),
-CLLocationCoordinate2D(latitude:47.7726138163672,longitude:-122.193985220722),
-CLLocationCoordinate2D(latitude:47.7726076688847,longitude:-122.194039228924),
-CLLocationCoordinate2D(latitude:47.7726052892141,longitude:-122.19409345847),
-CLLocationCoordinate2D(latitude:47.7726052892141,longitude:-122.194149680123),
-CLLocationCoordinate2D(latitude:47.7726052892141,longitude:-122.19420390967),
-CLLocationCoordinate2D(latitude:47.7726044959905,longitude:-122.194257917871),
-CLLocationCoordinate2D(latitude:47.7725937874726,longitude:-122.194309933967),
-CLLocationCoordinate2D(latitude:47.7725759399428,longitude:-122.194358187197),
-CLLocationCoordinate2D(latitude:47.7725541262953,longitude:-122.194399578728),
-CLLocationCoordinate2D(latitude:47.7725325109536,longitude:-122.194444511781),
-CLLocationCoordinate2D(latitude:47.7725130769767,longitude:-122.19449055156),
-CLLocationCoordinate2D(latitude:47.7724861073761,longitude:-122.194544781106),
-CLLocationCoordinate2D(latitude:47.7724666733992,longitude:-122.194590820885),
-CLLocationCoordinate2D(latitude:47.7724448597516,longitude:-122.194633319142),
-CLLocationCoordinate2D(latitude:47.772424830857,longitude:-122.194677145469),
-CLLocationCoordinate2D(latitude:47.7724053968801,longitude:-122.194724291973),
-CLLocationCoordinate2D(latitude:47.7724030172095,longitude:-122.19477719345),
-CLLocationCoordinate2D(latitude:47.7724030172095,longitude:-122.194830316271),
-CLLocationCoordinate2D(latitude:47.7724155104804,longitude:-122.194881004296),
-CLLocationCoordinate2D(latitude:47.7724349444573,longitude:-122.194937447294),
-CLLocationCoordinate2D(latitude:47.7724488258694,longitude:-122.194989242045),
-CLLocationCoordinate2D(latitude:47.7724619140579,longitude:-122.195038601999),
-CLLocationCoordinate2D(latitude:47.7724789683642,longitude:-122.195107661667),
-CLLocationCoordinate2D(latitude:47.772491461635,longitude:-122.195158349692),
-CLLocationCoordinate2D(latitude:47.7725053430471,longitude:-122.195207930992),
-CLLocationCoordinate2D(latitude:47.7725162498709,longitude:-122.195258619018),
-CLLocationCoordinate2D(latitude:47.7725200176827,longitude:-122.195321923713),
-CLLocationCoordinate2D(latitude:47.7725208109063,longitude:-122.195375931915),
-CLLocationCoordinate2D(latitude:47.7725216041298,longitude:-122.195429054736),
-CLLocationCoordinate2D(latitude:47.7725263634711,longitude:-122.195510731074),
-CLLocationCoordinate2D(latitude:47.7725340974007,longitude:-122.19556363255),
-CLLocationCoordinate2D(latitude:47.772537071989,longitude:-122.195616755371),
-CLLocationCoordinate2D(latitude:47.7725364770713,longitude:-122.195669656847),
-CLLocationCoordinate2D(latitude:47.7725279499182,longitude:-122.19572499312),
-CLLocationCoordinate2D(latitude:47.7725170430944,longitude:-122.19577678787),
-CLLocationCoordinate2D(latitude:47.7725069294942,longitude:-122.195827475896),
-CLLocationCoordinate2D(latitude:47.7724984023411,longitude:-122.195896535563),
-CLLocationCoordinate2D(latitude:47.772496815894,longitude:-122.19595165049),
-CLLocationCoordinate2D(latitude:47.7724946345292,longitude:-122.196006986763),
-CLLocationCoordinate2D(latitude:47.7724938413057,longitude:-122.196059888239),
-CLLocationCoordinate2D(latitude:47.7724976091175,longitude:-122.196122086209),
-CLLocationCoordinate2D(latitude:47.7725061362707,longitude:-122.19617388096),
-CLLocationCoordinate2D(latitude:47.7725146634238,longitude:-122.196227003781),
-],
-[
-CLLocationCoordinate2D(latitude:47.7741798379557,longitude:-122.193332252179),
-CLLocationCoordinate2D(latitude:47.7741830108499,longitude:-122.193278243978),
-CLLocationCoordinate2D(latitude:47.774189951556,longitude:-122.193222908767),
-CLLocationCoordinate2D(latitude:47.7741976854855,longitude:-122.193154955294),
-CLLocationCoordinate2D(latitude:47.7742040312739,longitude:-122.193100726278),
-CLLocationCoordinate2D(latitude:47.7742131533447,longitude:-122.193048930466),
-CLLocationCoordinate2D(latitude:47.7742232669449,longitude:-122.192997136777),
-CLLocationCoordinate2D(latitude:47.7742349669923,longitude:-122.192943128575),
-CLLocationCoordinate2D(latitude:47.7742480551808,longitude:-122.192893547275),
-CLLocationCoordinate2D(latitude:47.77426352304,longitude:-122.192843965975),
-CLLocationCoordinate2D(latitude:47.7742799824286,longitude:-122.192792170163),
-CLLocationCoordinate2D(latitude:47.7742946570642,longitude:-122.192741482668),
-CLLocationCoordinate2D(latitude:47.7743047706644,longitude:-122.192688580661),
-CLLocationCoordinate2D(latitude:47.7743101249234,longitude:-122.192635679716),
-CLLocationCoordinate2D(latitude:47.7743154791823,longitude:-122.192574588471),
-CLLocationCoordinate2D(latitude:47.7743232131119,longitude:-122.192512389971),
-],
-[
-CLLocationCoordinate2D(latitude:47.7743783421485,longitude:-122.191761809306),
-CLLocationCoordinate2D(latitude:47.7743714014424,longitude:-122.191707579228),
-CLLocationCoordinate2D(latitude:47.7743666421011,longitude:-122.191653571026),
-CLLocationCoordinate2D(latitude:47.7743604946187,longitude:-122.19159934201),
-CLLocationCoordinate2D(latitude:47.7743589081716,longitude:-122.191546440003),
-CLLocationCoordinate2D(latitude:47.7743589081716,longitude:-122.191493539058),
-CLLocationCoordinate2D(latitude:47.7743557352774,longitude:-122.191440416237),
-CLLocationCoordinate2D(latitude:47.7743434403124,longitude:-122.19138397377),
-CLLocationCoordinate2D(latitude:47.774345819983,longitude:-122.191313806846),
-],
-[
-CLLocationCoordinate2D(latitude:47.7748967137367,longitude:-122.194114043564),
-CLLocationCoordinate2D(latitude:47.7748967137367,longitude:-122.194061142087),
-CLLocationCoordinate2D(latitude:47.7748967137367,longitude:-122.194005805815),
-CLLocationCoordinate2D(latitude:47.7748990934074,longitude:-122.193952904339),
-CLLocationCoordinate2D(latitude:47.7749004815486,longitude:-122.193898674792),
-CLLocationCoordinate2D(latitude:47.7749036544427,longitude:-122.193845773316),
-CLLocationCoordinate2D(latitude:47.7749060341134,longitude:-122.193784682072),
-CLLocationCoordinate2D(latitude:47.7749060341134,longitude:-122.193731780595),
-CLLocationCoordinate2D(latitude:47.774908413784,longitude:-122.193678879119),
-CLLocationCoordinate2D(latitude:47.7749161477136,longitude:-122.193625977643),
-CLLocationCoordinate2D(latitude:47.7749222951961,longitude:-122.193563780204),
-CLLocationCoordinate2D(latitude:47.7749308223493,longitude:-122.193510657383),
-CLLocationCoordinate2D(latitude:47.7749409359495,longitude:-122.193459968826),
-CLLocationCoordinate2D(latitude:47.7749524376909,longitude:-122.193409502146),
-CLLocationCoordinate2D(latitude:47.7749665174089,longitude:-122.19335881359),
-CLLocationCoordinate2D(latitude:47.7749827784916,longitude:-122.193310339546),
-CLLocationCoordinate2D(latitude:47.7749990395743,longitude:-122.193260980122),
-CLLocationCoordinate2D(latitude:47.7750160938806,longitude:-122.193211398823),
-CLLocationCoordinate2D(latitude:47.7750438567048,longitude:-122.193156061489),
-CLLocationCoordinate2D(latitude:47.775068050023,longitude:-122.193116883939),
-CLLocationCoordinate2D(latitude:47.7750950196236,longitude:-122.193076599133),
-CLLocationCoordinate2D(latitude:47.7751291282361,longitude:-122.193022590931),
-CLLocationCoordinate2D(latitude:47.7751547096955,longitude:-122.192985626832),
-CLLocationCoordinate2D(latitude:47.7751802911549,longitude:-122.192947776291),
-CLLocationCoordinate2D(latitude:47.7752026997201,longitude:-122.192906163414),
-CLLocationCoordinate2D(latitude:47.775222133697,longitude:-122.192859017441),
-CLLocationCoordinate2D(latitude:47.7752368083327,longitude:-122.192809436141),
-CLLocationCoordinate2D(latitude:47.7752477151564,longitude:-122.192757640329),
-CLLocationCoordinate2D(latitude:47.775255449086,longitude:-122.192702525933),
-CLLocationCoordinate2D(latitude:47.7752623897921,longitude:-122.192646082404),
-CLLocationCoordinate2D(latitude:47.7752663559098,longitude:-122.192590747194),
-CLLocationCoordinate2D(latitude:47.7752663559098,longitude:-122.192536738992),
-CLLocationCoordinate2D(latitude:47.7752647694627,longitude:-122.192483616171),
-CLLocationCoordinate2D(latitude:47.775243154121,longitude:-122.192438682587),
-CLLocationCoordinate2D(latitude:47.7752213404735,longitude:-122.192395078135),
-CLLocationCoordinate2D(latitude:47.7751981386847,longitude:-122.192353465259),
-CLLocationCoordinate2D(latitude:47.7751685911076,longitude:-122.192292594828),
-CLLocationCoordinate2D(latitude:47.7751648232957,longitude:-122.192238365812),
-CLLocationCoordinate2D(latitude:47.7751594690368,longitude:-122.192185463805),
-CLLocationCoordinate2D(latitude:47.7751493554366,longitude:-122.192133669055),
-CLLocationCoordinate2D(latitude:47.7751307146832,longitude:-122.192070364359),
-CLLocationCoordinate2D(latitude:47.7751112807063,longitude:-122.192024324581),
-CLLocationCoordinate2D(latitude:47.7750904585882,longitude:-122.191979390997),
-CLLocationCoordinate2D(latitude:47.7750640839052,longitude:-122.191921841274),
-CLLocationCoordinate2D(latitude:47.7750438567048,longitude:-122.191875801496),
-CLLocationCoordinate2D(latitude:47.7750198616924,longitude:-122.191835517751),
-CLLocationCoordinate2D(latitude:47.7749982463508,longitude:-122.191789477972),
-CLLocationCoordinate2D(latitude:47.7749726648914,longitude:-122.191731706374),
-CLLocationCoordinate2D(latitude:47.7749556105851,longitude:-122.191683453144),
-CLLocationCoordinate2D(latitude:47.7749377630553,longitude:-122.191637413366),
-CLLocationCoordinate2D(latitude:47.7749191223019,longitude:-122.191592479782),
-CLLocationCoordinate2D(latitude:47.7748897730307,longitude:-122.191533823864),
-CLLocationCoordinate2D(latitude:47.7748687526067,longitude:-122.191489997537),
-CLLocationCoordinate2D(latitude:47.774841783006,longitude:-122.191451926181),
-CLLocationCoordinate2D(latitude:47.7747898268637,longitude:-122.191401237625),
-],
-[
-CLLocationCoordinate2D(latitude:47.7756080369529,longitude:-122.191610851955),
-CLLocationCoordinate2D(latitude:47.7756227115885,longitude:-122.19165932706),
-CLLocationCoordinate2D(latitude:47.775623504812,longitude:-122.191712228006),
-CLLocationCoordinate2D(latitude:47.7756088301764,longitude:-122.191769777729),
-CLLocationCoordinate2D(latitude:47.7755909826466,longitude:-122.191816924764),
-CLLocationCoordinate2D(latitude:47.775570953752,longitude:-122.191862964542),
-CLLocationCoordinate2D(latitude:47.7755515197751,longitude:-122.191907897064),
-CLLocationCoordinate2D(latitude:47.775528912904,longitude:-122.191950617197),
-CLLocationCoordinate2D(latitude:47.7755104704565,longitude:-122.191998869365),
-CLLocationCoordinate2D(latitude:47.7754902432561,longitude:-122.192044909143),
-CLLocationCoordinate2D(latitude:47.7754638685731,longitude:-122.192080768109),
-CLLocationCoordinate2D(latitude:47.7754523668317,longitude:-122.19213123479),
-CLLocationCoordinate2D(latitude:47.7754507803846,longitude:-122.192184357611),
-CLLocationCoordinate2D(latitude:47.7754545481964,longitude:-122.192283298335),
-CLLocationCoordinate2D(latitude:47.7754561346435,longitude:-122.192338634607),
-CLLocationCoordinate2D(latitude:47.7754638685731,longitude:-122.192390429358),
-CLLocationCoordinate2D(latitude:47.775476361844,longitude:-122.192441117914),
-CLLocationCoordinate2D(latitude:47.7754934161503,longitude:-122.192487157692),
-CLLocationCoordinate2D(latitude:47.7755142382684,longitude:-122.192537845187),
-CLLocationCoordinate2D(latitude:47.7755336722453,longitude:-122.192586098416),
-CLLocationCoordinate2D(latitude:47.7755568740341,longitude:-122.192627711293),
-CLLocationCoordinate2D(latitude:47.7755816622699,longitude:-122.192678177973),
-CLLocationCoordinate2D(latitude:47.7756026826939,longitude:-122.192722004301),
-CLLocationCoordinate2D(latitude:47.7756266777062,longitude:-122.192763396363),
-CLLocationCoordinate2D(latitude:47.7756506727185,longitude:-122.192801467718),
-CLLocationCoordinate2D(latitude:47.7756847813311,longitude:-122.19283378304),
-CLLocationCoordinate2D(latitude:47.7757157170494,longitude:-122.192861229831),
-CLLocationCoordinate2D(latitude:47.775744471403,longitude:-122.19289598154),
-CLLocationCoordinate2D(latitude:47.775770846086,longitude:-122.19293161863),
-CLLocationCoordinate2D(latitude:47.7758049546985,longitude:-122.19298341232),
-CLLocationCoordinate2D(latitude:47.7758249835931,longitude:-122.193027239709),
-CLLocationCoordinate2D(latitude:47.7758428311229,longitude:-122.193074385682),
-CLLocationCoordinate2D(latitude:47.7758699990294,longitude:-122.19316867869),
-CLLocationCoordinate2D(latitude:47.775884673665,longitude:-122.19321825999),
-CLLocationCoordinate2D(latitude:47.7758971669359,longitude:-122.193268948546),
-CLLocationCoordinate2D(latitude:47.7759094619009,longitude:-122.193320742235),
-CLLocationCoordinate2D(latitude:47.7759211619482,longitude:-122.193374972312),
-CLLocationCoordinate2D(latitude:47.7759288958778,longitude:-122.193427873258),
-CLLocationCoordinate2D(latitude:47.775939009478,longitude:-122.193478561814),
-CLLocationCoordinate2D(latitude:47.7759475366312,longitude:-122.193530356565),
-],
-[
-CLLocationCoordinate2D(latitude:47.7758753532884,longitude:-122.196649551555),
-CLLocationCoordinate2D(latitude:47.7758606786527,longitude:-122.196642689857),
-CLLocationCoordinate2D(latitude:47.7758715854765,longitude:-122.196593108558),
-CLLocationCoordinate2D(latitude:47.7758932008182,longitude:-122.196543527258),
-CLLocationCoordinate2D(latitude:47.7759156093834,longitude:-122.196502135726),
-CLLocationCoordinate2D(latitude:47.775944363737,longitude:-122.196471147414),
-CLLocationCoordinate2D(latitude:47.775977679126,longitude:-122.196451447701),
-CLLocationCoordinate2D(latitude:47.7760304284919,longitude:-122.196431969333),
-CLLocationCoordinate2D(latitude:47.7760659252456,longitude:-122.196419131318),
-CLLocationCoordinate2D(latitude:47.7761008270817,longitude:-122.196401866401),
-CLLocationCoordinate2D(latitude:47.7761341424707,longitude:-122.196380174582),
-CLLocationCoordinate2D(latitude:47.7761706307539,longitude:-122.196352506446),
-CLLocationCoordinate2D(latitude:47.7761985918839,longitude:-122.196319083338),
-CLLocationCoordinate2D(latitude:47.7762225868962,longitude:-122.196279905257),
-CLLocationCoordinate2D(latitude:47.7762348818612,longitude:-122.196230323957),
-CLLocationCoordinate2D(latitude:47.7762418225673,longitude:-122.196154402592),
-],
-[
-CLLocationCoordinate2D(latitude:47.7762681972502,longitude:-122.195921768903),
-CLLocationCoordinate2D(latitude:47.776279104074,longitude:-122.195871080878),
-CLLocationCoordinate2D(latitude:47.7762836651094,longitude:-122.195818179402),
-CLLocationCoordinate2D(latitude:47.7762836651094,longitude:-122.195765277926),
-CLLocationCoordinate2D(latitude:47.7762814837446,longitude:-122.195711048379),
-CLLocationCoordinate2D(latitude:47.7762836651094,longitude:-122.195658146903),
-CLLocationCoordinate2D(latitude:47.7762921922625,longitude:-122.195605245427),
-CLLocationCoordinate2D(latitude:47.7763023058628,longitude:-122.195537292484),
-CLLocationCoordinate2D(latitude:47.7763108330159,longitude:-122.195485497734),
-CLLocationCoordinate2D(latitude:47.7763185669455,longitude:-122.195433702983),
-CLLocationCoordinate2D(latitude:47.7763255076515,longitude:-122.195378366711),
-CLLocationCoordinate2D(latitude:47.7763340348047,longitude:-122.195327678685),
-CLLocationCoordinate2D(latitude:47.7763419670401,longitude:-122.195270128962),
-CLLocationCoordinate2D(latitude:47.776345734852,longitude:-122.195215899415),
-CLLocationCoordinate2D(latitude:47.7763497009697,longitude:-122.195162997939),
-CLLocationCoordinate2D(latitude:47.7763518823345,longitude:-122.195106554942),
-CLLocationCoordinate2D(latitude:47.7763518823345,longitude:-122.195053653465),
-CLLocationCoordinate2D(latitude:47.7763449416284,longitude:-122.195001858715),
-CLLocationCoordinate2D(latitude:47.7763263008751,longitude:-122.194956925662),
-CLLocationCoordinate2D(latitude:47.7763140059101,longitude:-122.194901589389),
-CLLocationCoordinate2D(latitude:47.7763054787569,longitude:-122.194849794639),
-CLLocationCoordinate2D(latitude:47.7763046855334,longitude:-122.194796893162),
-CLLocationCoordinate2D(latitude:47.7763177737219,longitude:-122.194743991686),
-CLLocationCoordinate2D(latitude:47.7763278873222,longitude:-122.194693303661),
-CLLocationCoordinate2D(latitude:47.7763364144753,longitude:-122.194639074114),
-CLLocationCoordinate2D(latitude:47.7763441484049,longitude:-122.194587279363),
-CLLocationCoordinate2D(latitude:47.776352675558,longitude:-122.19451711297),
-CLLocationCoordinate2D(latitude:47.7763705230878,longitude:-122.194471073192),
-CLLocationCoordinate2D(latitude:47.7763867841706,longitude:-122.194422598617),
-CLLocationCoordinate2D(latitude:47.7763992774414,longitude:-122.194370803867),
-CLLocationCoordinate2D(latitude:47.7764099859593,longitude:-122.194299530748),
-CLLocationCoordinate2D(latitude:47.7764171249712,longitude:-122.194246407927),
-CLLocationCoordinate2D(latitude:47.7764248589008,longitude:-122.194194613176),
-CLLocationCoordinate2D(latitude:47.7764355674187,longitude:-122.194143925151),
-CLLocationCoordinate2D(latitude:47.7764496471367,longitude:-122.194094565196),
-CLLocationCoordinate2D(latitude:47.776466701443,longitude:-122.194048525417),
-CLLocationCoordinate2D(latitude:47.7764944642671,longitude:-122.193983892651),
-CLLocationCoordinate2D(latitude:47.7765146914676,longitude:-122.193938959598),
-CLLocationCoordinate2D(latitude:47.7765378932563,longitude:-122.193897568067),
-CLLocationCoordinate2D(latitude:47.7765610950451,longitude:-122.193857283261),
-CLLocationCoordinate2D(latitude:47.7765866765045,longitude:-122.193816998455),
-CLLocationCoordinate2D(latitude:47.7766239580112,longitude:-122.193788223593),
-CLLocationCoordinate2D(latitude:47.7766604462944,longitude:-122.193784682072),
-CLLocationCoordinate2D(latitude:47.7766959430482,longitude:-122.193778927099),
-CLLocationCoordinate2D(latitude:47.7767494856376,longitude:-122.193766310429),
-CLLocationCoordinate2D(latitude:47.7767859739208,longitude:-122.193747938787),
-CLLocationCoordinate2D(latitude:47.7768184960862,longitude:-122.193724918898),
-CLLocationCoordinate2D(latitude:47.7768531996164,longitude:-122.193707653981),
-CLLocationCoordinate2D(latitude:47.7768904811232,longitude:-122.193689282338),
-CLLocationCoordinate2D(latitude:47.7769253829593,longitude:-122.193674230872),
-CLLocationCoordinate2D(latitude:47.7769602847953,longitude:-122.193661613671),
-CLLocationCoordinate2D(latitude:47.7770058951493,longitude:-122.193627083838),
-CLLocationCoordinate2D(latitude:47.7770392105383,longitude:-122.193595874711),
-CLLocationCoordinate2D(latitude:47.777069551339,longitude:-122.193565992593),
-CLLocationCoordinate2D(latitude:47.7770989006103,longitude:-122.193531462759),
-CLLocationCoordinate2D(latitude:47.7771260685168,longitude:-122.193494720536),
-CLLocationCoordinate2D(latitude:47.7771492703055,longitude:-122.193454213854),
-CLLocationCoordinate2D(latitude:47.7771716788707,longitude:-122.193413929048),
-CLLocationCoordinate2D(latitude:47.7771919060712,longitude:-122.193366782013),
-CLLocationCoordinate2D(latitude:47.7772057874833,longitude:-122.193318529846),
-CLLocationCoordinate2D(latitude:47.7772337486133,longitude:-122.193227556483),
-CLLocationCoordinate2D(latitude:47.7772470351077,longitude:-122.193176868989),
-CLLocationCoordinate2D(latitude:47.7772577436256,longitude:-122.193126180433),
-CLLocationCoordinate2D(latitude:47.7772609165198,longitude:-122.193071950355),
-CLLocationCoordinate2D(latitude:47.7772787640496,longitude:-122.193022590931),
-CLLocationCoordinate2D(latitude:47.7773065268738,longitude:-122.192988061097),
-CLLocationCoordinate2D(latitude:47.7773190201447,longitude:-122.192938479797),
-CLLocationCoordinate2D(latitude:47.7773198133682,longitude:-122.192885578852),
-CLLocationCoordinate2D(latitude:47.7773104929915,longitude:-122.192816519184),
-CLLocationCoordinate2D(latitude:47.7772894725675,longitude:-122.192773799051),
-CLLocationCoordinate2D(latitude:47.7772779708261,longitude:-122.192724217751),
-CLLocationCoordinate2D(latitude:47.7772942319088,longitude:-122.192654050828),
-CLLocationCoordinate2D(latitude:47.777304345509,longitude:-122.192602257138),
-CLLocationCoordinate2D(latitude:47.7773142608034,longitude:-122.192551568582),
-CLLocationCoordinate2D(latitude:47.7773182269211,longitude:-122.192497560381),
-CLLocationCoordinate2D(latitude:47.7773174336976,longitude:-122.192444437559),
-CLLocationCoordinate2D(latitude:47.7772942319088,longitude:-122.192404152753),
-CLLocationCoordinate2D(latitude:47.7772724182613,longitude:-122.192363867947),
-CLLocationCoordinate2D(latitude:47.77724306899,longitude:-122.192333986891),
-CLLocationCoordinate2D(latitude:47.7772137197187,longitude:-122.192302998578),
-CLLocationCoordinate2D(latitude:47.7771825856945,longitude:-122.192263819967),
-],
-[
-CLLocationCoordinate2D(latitude:47.7764936710436,longitude:-122.196572302119),
-CLLocationCoordinate2D(latitude:47.7764821693021,longitude:-122.196520507368),
-CLLocationCoordinate2D(latitude:47.7764696760313,longitude:-122.196469819343),
-CLLocationCoordinate2D(latitude:47.7764573810663,longitude:-122.196415811142),
-CLLocationCoordinate2D(latitude:47.7764440945719,longitude:-122.196365123116),
-CLLocationCoordinate2D(latitude:47.7764294199362,longitude:-122.196316648542),
-CLLocationCoordinate2D(latitude:47.7764177198889,longitude:-122.196264853791),
-CLLocationCoordinate2D(latitude:47.77641236563,longitude:-122.196209738864),
-CLLocationCoordinate2D(latitude:47.7764030452533,longitude:-122.196157944113),
-CLLocationCoordinate2D(latitude:47.7763907502883,longitude:-122.196104821292),
-CLLocationCoordinate2D(latitude:47.7763744892056,longitude:-122.196055461337),
-CLLocationCoordinate2D(latitude:47.7763550552287,longitude:-122.196009421559),
-CLLocationCoordinate2D(latitude:47.7763348280282,longitude:-122.195964488506),
-CLLocationCoordinate2D(latitude:47.7763140059101,longitude:-122.195920662178),
-CLLocationCoordinate2D(latitude:47.7762884244507,longitude:-122.195850495785),
-],
-[
-CLLocationCoordinate2D(latitude:47.7772313689427,longitude:-122.195092831546),
-CLLocationCoordinate2D(latitude:47.7772531825902,longitude:-122.195050111944),
-CLLocationCoordinate2D(latitude:47.7772809454144,longitude:-122.194992562221),
-CLLocationCoordinate2D(latitude:47.7773003793913,longitude:-122.194944308992),
-CLLocationCoordinate2D(latitude:47.7773174336976,longitude:-122.194895834417),
-CLLocationCoordinate2D(latitude:47.7773336947803,longitude:-122.194846474462),
-CLLocationCoordinate2D(latitude:47.7773515423101,longitude:-122.194799106613),
-CLLocationCoordinate2D(latitude:47.7773717695106,longitude:-122.194747311863),
-CLLocationCoordinate2D(latitude:47.7773872373697,longitude:-122.194697951908),
-CLLocationCoordinate2D(latitude:47.7774011187818,longitude:-122.194648370608),
-CLLocationCoordinate2D(latitude:47.777416586641,longitude:-122.194593034336),
-CLLocationCoordinate2D(latitude:47.7774384002886,longitude:-122.194535484613),
-CLLocationCoordinate2D(latitude:47.7774562478184,longitude:-122.194488338109),
-CLLocationCoordinate2D(latitude:47.77747786316,longitude:-122.194445618507),
-CLLocationCoordinate2D(latitude:47.7775080056548,longitude:-122.194393823756),
-CLLocationCoordinate2D(latitude:47.7775290260788,longitude:-122.194351325499),
-CLLocationCoordinate2D(latitude:47.7775484600558,longitude:-122.19430528572),
-CLLocationCoordinate2D(latitude:47.7775639279149,longitude:-122.194257917871),
-CLLocationCoordinate2D(latitude:47.7775770161035,longitude:-122.19420390967),
-CLLocationCoordinate2D(latitude:47.7775855432566,longitude:-122.194152114919),
-CLLocationCoordinate2D(latitude:47.7775932771862,longitude:-122.194099213443),
-CLLocationCoordinate2D(latitude:47.7775948636333,longitude:-122.194046090621),
-CLLocationCoordinate2D(latitude:47.7775924839626,longitude:-122.193990975694),
-CLLocationCoordinate2D(latitude:47.7775746364328,longitude:-122.193943607845),
-CLLocationCoordinate2D(latitude:47.7775506414205,longitude:-122.193902216314),
-CLLocationCoordinate2D(latitude:47.7775220853728,longitude:-122.193869899931),
-CLLocationCoordinate2D(latitude:47.7775127649961,longitude:-122.193819211905),
-CLLocationCoordinate2D(latitude:47.777532198973,longitude:-122.193776713648),
-CLLocationCoordinate2D(latitude:47.7775538143147,longitude:-122.193732887321),
-CLLocationCoordinate2D(latitude:47.777570868621,longitude:-122.193676444323),
-CLLocationCoordinate2D(latitude:47.7775823703624,longitude:-122.193623542316),
-CLLocationCoordinate2D(latitude:47.7775924839626,longitude:-122.193570641371),
-CLLocationCoordinate2D(latitude:47.7776010111158,longitude:-122.193517740425),
-CLLocationCoordinate2D(latitude:47.7776049772335,longitude:-122.193463510348),
-CLLocationCoordinate2D(latitude:47.7776057704571,longitude:-122.193410609402),
-CLLocationCoordinate2D(latitude:47.7775863364802,longitude:-122.193364569624),
-CLLocationCoordinate2D(latitude:47.7775647211385,longitude:-122.193311446803),
-CLLocationCoordinate2D(latitude:47.7775490549734,longitude:-122.193263192512),
-CLLocationCoordinate2D(latitude:47.7775290260788,longitude:-122.19321825999),
-CLLocationCoordinate2D(latitude:47.7775080056548,longitude:-122.193156061489),
-CLLocationCoordinate2D(latitude:47.7774980903605,longitude:-122.1931042678),
-CLLocationCoordinate2D(latitude:47.7774925377957,longitude:-122.193052471987),
-CLLocationCoordinate2D(latitude:47.7774965039134,longitude:-122.192999571042),
-CLLocationCoordinate2D(latitude:47.7774972971369,longitude:-122.19293383102),
-CLLocationCoordinate2D(latitude:47.7774794496071,longitude:-122.192886685047),
-CLLocationCoordinate2D(latitude:47.7774608088538,longitude:-122.192840645268),
-CLLocationCoordinate2D(latitude:47.7774368138415,longitude:-122.192785308996),
-CLLocationCoordinate2D(latitude:47.7774368138415,longitude:-122.192731300794),
-CLLocationCoordinate2D(latitude:47.7774405816533,longitude:-122.192678177973),
-CLLocationCoordinate2D(latitude:47.7774546613713,longitude:-122.192628818549),
-CLLocationCoordinate2D(latitude:47.7774770699365,longitude:-122.192581449638),
-CLLocationCoordinate2D(latitude:47.777511178549,longitude:-122.192529655949),
-],
-[
-CLLocationCoordinate2D(latitude:47.7783434683562,longitude:-122.195752661256),
-CLLocationCoordinate2D(latitude:47.7783232411557,longitude:-122.195707728203),
-CLLocationCoordinate2D(latitude:47.778313920779,longitude:-122.195654605381),
-CLLocationCoordinate2D(latitude:47.7783115411084,longitude:-122.195601703905),
-CLLocationCoordinate2D(latitude:47.7782976596963,longitude:-122.195554557401),
-CLLocationCoordinate2D(latitude:47.778273664684,longitude:-122.195514272595),
-CLLocationCoordinate2D(latitude:47.7782457035539,longitude:-122.195478414691),
-CLLocationCoordinate2D(latitude:47.7782248814358,longitude:-122.195432374912),
-CLLocationCoordinate2D(latitude:47.7782102068002,longitude:-122.195379473436),
-CLLocationCoordinate2D(latitude:47.7782038610118,longitude:-122.195325465234),
-CLLocationCoordinate2D(latitude:47.7782038610118,longitude:-122.195272342413),
-CLLocationCoordinate2D(latitude:47.7782179407298,longitude:-122.195195314322),
-CLLocationCoordinate2D(latitude:47.7782326153654,longitude:-122.195145733022),
-CLLocationCoordinate2D(latitude:47.7782512561188,longitude:-122.195100799969),
-CLLocationCoordinate2D(latitude:47.7782683104251,longitude:-122.195053653465),
-CLLocationCoordinate2D(latitude:47.7782720782369,longitude:-122.195000751989),
-CLLocationCoordinate2D(latitude:47.7782696985663,longitude:-122.194946522442),
-CLLocationCoordinate2D(latitude:47.7782589900484,longitude:-122.194894727692),
-CLLocationCoordinate2D(latitude:47.7782387628479,longitude:-122.194819913052),
-CLLocationCoordinate2D(latitude:47.7782272611065,longitude:-122.194767011575),
-CLLocationCoordinate2D(latitude:47.7782155610591,longitude:-122.19471632355),
-CLLocationCoordinate2D(latitude:47.7782046542354,longitude:-122.194660987278),
-CLLocationCoordinate2D(latitude:47.7781939457175,longitude:-122.194589714159),
-CLLocationCoordinate2D(latitude:47.7781891863762,longitude:-122.194535484613),
-CLLocationCoordinate2D(latitude:47.7781875999291,longitude:-122.194482583136),
-CLLocationCoordinate2D(latitude:47.7781875999291,longitude:-122.19442835359),
-CLLocationCoordinate2D(latitude:47.7781939457175,longitude:-122.194376558839),
-CLLocationCoordinate2D(latitude:47.7782115949414,longitude:-122.194312147418),
-CLLocationCoordinate2D(latitude:47.778233408589,longitude:-122.194269427816),
-CLLocationCoordinate2D(latitude:47.7782605764955,longitude:-122.19423047108),
-CLLocationCoordinate2D(latitude:47.7783032122611,longitude:-122.194187751478),
-CLLocationCoordinate2D(latitude:47.7783365276501,longitude:-122.19416694504),
-CLLocationCoordinate2D(latitude:47.7783720244039,longitude:-122.194151008194),
-CLLocationCoordinate2D(latitude:47.7784053397929,longitude:-122.194126660234),
-CLLocationCoordinate2D(latitude:47.7784450009702,longitude:-122.194088810224),
-CLLocationCoordinate2D(latitude:47.7784650298648,longitude:-122.194033473951),
-],
-[
-CLLocationCoordinate2D(latitude:47.7791493834804,longitude:-122.194670283772),
-CLLocationCoordinate2D(latitude:47.7791664377866,longitude:-122.194622915923),
-CLLocationCoordinate2D(latitude:47.7791856734577,longitude:-122.194578204215),
-CLLocationCoordinate2D(latitude:47.7792051074346,longitude:-122.194530836366),
-CLLocationCoordinate2D(latitude:47.7792245414115,longitude:-122.194479041615),
-CLLocationCoordinate2D(latitude:47.7792376296,longitude:-122.19442968166),
-CLLocationCoordinate2D(latitude:47.7792453635296,longitude:-122.194376558839),
-CLLocationCoordinate2D(latitude:47.7792493296473,longitude:-122.194323657363),
-CLLocationCoordinate2D(latitude:47.7792515110121,longitude:-122.194268321091),
-CLLocationCoordinate2D(latitude:47.7792515110121,longitude:-122.194214312889),
-CLLocationCoordinate2D(latitude:47.7792515110121,longitude:-122.194160083342),
-CLLocationCoordinate2D(latitude:47.7792493296473,longitude:-122.194107181866),
-CLLocationCoordinate2D(latitude:47.7792408024942,longitude:-122.194033473951),
-CLLocationCoordinate2D(latitude:47.7792360431529,longitude:-122.193977030954),
-CLLocationCoordinate2D(latitude:47.7792322753411,longitude:-122.193924129478),
-CLLocationCoordinate2D(latitude:47.7792283092233,longitude:-122.193869899931),
-CLLocationCoordinate2D(latitude:47.7792197820702,longitude:-122.193813456933),
-CLLocationCoordinate2D(latitude:47.7792104616935,longitude:-122.193759448732),
-CLLocationCoordinate2D(latitude:47.7791981667285,longitude:-122.19370544053),
-CLLocationCoordinate2D(latitude:47.7791787327516,longitude:-122.193639701038),
-CLLocationCoordinate2D(latitude:47.779170998822,longitude:-122.193587906288),
-CLLocationCoordinate2D(latitude:47.7791632648924,longitude:-122.193536111537),
-CLLocationCoordinate2D(latitude:47.7791585055511,longitude:-122.193483210592),
-],
-[
-CLLocationCoordinate2D(latitude:47.7796661686215,longitude:-122.194386962058),
-CLLocationCoordinate2D(latitude:47.779666961845,longitude:-122.194334060582),
-CLLocationCoordinate2D(latitude:47.7796693415157,longitude:-122.194280937761),
-CLLocationCoordinate2D(latitude:47.7796693415157,longitude:-122.194228036284),
-CLLocationCoordinate2D(latitude:47.7796693415157,longitude:-122.194175134808),
-CLLocationCoordinate2D(latitude:47.7796746957746,longitude:-122.194123340057),
-CLLocationCoordinate2D(latitude:47.7796856025984,longitude:-122.194072652032),
-CLLocationCoordinate2D(latitude:47.7796986907869,longitude:-122.194023070732),
-CLLocationCoordinate2D(latitude:47.7797141586461,longitude:-122.193972382707),
-CLLocationCoordinate2D(latitude:47.7797234790228,longitude:-122.193920587956),
-CLLocationCoordinate2D(latitude:47.7797242722464,longitude:-122.193865473029),
-CLLocationCoordinate2D(latitude:47.779719711211,longitude:-122.19380548851),
-CLLocationCoordinate2D(latitude:47.7797103908343,longitude:-122.193754800485),
-CLLocationCoordinate2D(latitude:47.7796955178928,longitude:-122.193704112459),
-],
-[
-CLLocationCoordinate2D(latitude:47.7800366040182,longitude:-122.196453661152),
-CLLocationCoordinate2D(latitude:47.7800536583244,longitude:-122.196406514648),
-CLLocationCoordinate2D(latitude:47.7800691261836,longitude:-122.196358261418),
-CLLocationCoordinate2D(latitude:47.780080627925,longitude:-122.196301818421),
-CLLocationCoordinate2D(latitude:47.780080627925,longitude:-122.196248916945),
-CLLocationCoordinate2D(latitude:47.780073687219,longitude:-122.196195794123),
-CLLocationCoordinate2D(latitude:47.7800721007719,longitude:-122.196142892647),
-CLLocationCoordinate2D(latitude:47.7800768601132,longitude:-122.196070291458),
-CLLocationCoordinate2D(latitude:47.7800869737134,longitude:-122.196018496707),
-CLLocationCoordinate2D(latitude:47.7801000619019,longitude:-122.195967808682),
-CLLocationCoordinate2D(latitude:47.7801218755495,longitude:-122.195894100767),
-CLLocationCoordinate2D(latitude:47.7801373434087,longitude:-122.195842306017),
-CLLocationCoordinate2D(latitude:47.7801520180443,longitude:-122.195791617991),
-CLLocationCoordinate2D(latitude:47.7801651062328,longitude:-122.195739823241),
-CLLocationCoordinate2D(latitude:47.7801922741393,longitude:-122.195669656847),
-CLLocationCoordinate2D(latitude:47.7802117081162,longitude:-122.19562583052),
-CLLocationCoordinate2D(latitude:47.7802255895283,longitude:-122.195575142494),
-CLLocationCoordinate2D(latitude:47.7802317370108,longitude:-122.195522241018),
-CLLocationCoordinate2D(latitude:47.7802450235052,longitude:-122.195458936323),
-CLLocationCoordinate2D(latitude:47.7802658456234,longitude:-122.195415109995),
-CLLocationCoordinate2D(latitude:47.7802844863767,longitude:-122.195369070217),
-CLLocationCoordinate2D(latitude:47.7803063000243,longitude:-122.195293148851),
-CLLocationCoordinate2D(latitude:47.7803178017657,longitude:-122.195240247375),
-CLLocationCoordinate2D(latitude:47.780329501813,longitude:-122.195188452624),
-CLLocationCoordinate2D(latitude:47.7803396154133,longitude:-122.195120499682),
-CLLocationCoordinate2D(latitude:47.7803441764487,longitude:-122.195061621889),
-CLLocationCoordinate2D(latitude:47.7803473493429,longitude:-122.195006506961),
-CLLocationCoordinate2D(latitude:47.7803580578607,longitude:-122.194955818936),
-CLLocationCoordinate2D(latitude:47.7803852257672,longitude:-122.19490291746),
-CLLocationCoordinate2D(latitude:47.7804153682621,longitude:-122.194871707802),
-CLLocationCoordinate2D(latitude:47.7804472955098,longitude:-122.194842932941),
-CLLocationCoordinate2D(latitude:47.7804790244517,longitude:-122.194809509833),
-CLLocationCoordinate2D(latitude:47.7805123398407,longitude:-122.194776086724),
-CLLocationCoordinate2D(latitude:47.7805424823355,longitude:-122.194743991686),
-CLLocationCoordinate2D(latitude:47.7805728231362,longitude:-122.194712782029),
-CLLocationCoordinate2D(latitude:47.7806184334902,longitude:-122.194670283772),
-CLLocationCoordinate2D(latitude:47.7806501624321,longitude:-122.19464018084),
-],
-[
-CLLocationCoordinate2D(latitude:47.7823752253419,longitude:-122.198085417146),
-CLLocationCoordinate2D(latitude:47.7823651117416,longitude:-122.198034729121),
-CLLocationCoordinate2D(latitude:47.7823581710356,longitude:-122.197981606299),
-CLLocationCoordinate2D(latitude:47.7823464709883,longitude:-122.197929811549),
-CLLocationCoordinate2D(latitude:47.7823317963526,longitude:-122.197881558319),
-CLLocationCoordinate2D(latitude:47.7823139488228,longitude:-122.197835518541),
-CLLocationCoordinate2D(latitude:47.7822867809163,longitude:-122.197798553911),
-CLLocationCoordinate2D(latitude:47.782250490939,longitude:-122.197761810626),
-CLLocationCoordinate2D(latitude:47.7822177704677,longitude:-122.197736355941),
-CLLocationCoordinate2D(latitude:47.7821836618552,longitude:-122.197714664122),
-CLLocationCoordinate2D(latitude:47.7821334904658,longitude:-122.197701826107),
-CLLocationCoordinate2D(latitude:47.7820977954061,longitude:-122.197688102712),
-CLLocationCoordinate2D(latitude:47.7820605138994,longitude:-122.197678806218),
-CLLocationCoordinate2D(latitude:47.782015696769,longitude:-122.197674379316),
-CLLocationCoordinate2D(latitude:47.7819730610033,longitude:-122.197668624344),
-CLLocationCoordinate2D(latitude:47.7819373659437,longitude:-122.197663976097),
-CLLocationCoordinate2D(latitude:47.781901670884,longitude:-122.197661541301),
-CLLocationCoordinate2D(latitude:47.7818560605301,longitude:-122.197654679603),
-CLLocationCoordinate2D(latitude:47.7818187790233,longitude:-122.197651138082),
-CLLocationCoordinate2D(latitude:47.7817822907401,longitude:-122.19764671118),
-CLLocationCoordinate2D(latitude:47.7817467939864,longitude:-122.197638521412),
-CLLocationCoordinate2D(latitude:47.7817103057032,longitude:-122.197613288072),
-CLLocationCoordinate2D(latitude:47.7816785767613,longitude:-122.197587833386),
-CLLocationCoordinate2D(latitude:47.7816490291842,longitude:-122.197557951799),
-CLLocationCoordinate2D(latitude:47.7816226545012,longitude:-122.197521208515),
-CLLocationCoordinate2D(latitude:47.7815887441946,longitude:-122.197462330721),
-CLLocationCoordinate2D(latitude:47.7815693102177,longitude:-122.197416290943),
-CLLocationCoordinate2D(latitude:47.7815530491349,longitude:-122.197369144439),
-CLLocationCoordinate2D(latitude:47.7815328219345,longitude:-122.197313808167),
-CLLocationCoordinate2D(latitude:47.7815219151107,longitude:-122.197263120141),
-CLLocationCoordinate2D(latitude:47.7815242947814,longitude:-122.19720911194),
-CLLocationCoordinate2D(latitude:47.781533615158,longitude:-122.197156210463),
-CLLocationCoordinate2D(latitude:47.7815359948287,longitude:-122.19709866074),
-CLLocationCoordinate2D(latitude:47.7815274676755,longitude:-122.197046644644),
-CLLocationCoordinate2D(latitude:47.7815127930399,longitude:-122.19699728469),
-CLLocationCoordinate2D(latitude:47.781493359063,longitude:-122.196953458362),
-CLLocationCoordinate2D(latitude:47.78146698438,longitude:-122.196915387007),
-CLLocationCoordinate2D(latitude:47.7814421961442,longitude:-122.196873995475),
-CLLocationCoordinate2D(latitude:47.781419787579,longitude:-122.196832603944),
-CLLocationCoordinate2D(latitude:47.7813963874843,longitude:-122.196789884342),
-CLLocationCoordinate2D(latitude:47.7813670382131,longitude:-122.196742737838),
-CLLocationCoordinate2D(latitude:47.7813414567537,longitude:-122.196704887827),
-CLLocationCoordinate2D(latitude:47.7813212295532,longitude:-122.196659954774),
-CLLocationCoordinate2D(latitude:47.7813134956236,longitude:-122.196608160024),
-CLLocationCoordinate2D(latitude:47.7813121074824,longitude:-122.196553930477),
-CLLocationCoordinate2D(latitude:47.7813089345882,longitude:-122.196501029001),
-CLLocationCoordinate2D(latitude:47.7813004074351,longitude:-122.196450340975),
-CLLocationCoordinate2D(latitude:47.7812841463524,longitude:-122.196398546225),
-CLLocationCoordinate2D(latitude:47.7812577716694,longitude:-122.196362909665),
-CLLocationCoordinate2D(latitude:47.7812516241869,longitude:-122.196310893569),
-CLLocationCoordinate2D(latitude:47.7812516241869,longitude:-122.196257992093),
-CLLocationCoordinate2D(latitude:47.7812468648456,longitude:-122.196203983892),
-CLLocationCoordinate2D(latitude:47.7812361563277,longitude:-122.196153295866),
-CLLocationCoordinate2D(latitude:47.7812191020215,longitude:-122.196103714566),
-CLLocationCoordinate2D(latitude:47.7812004612681,longitude:-122.196056568062),
-CLLocationCoordinate2D(latitude:47.7811725001381,longitude:-122.196018496707),
-CLLocationCoordinate2D(latitude:47.7811453322316,longitude:-122.195983966873),
-CLLocationCoordinate2D(latitude:47.7811191558545,longitude:-122.195942575342),
-CLLocationCoordinate2D(latitude:47.7810997218776,longitude:-122.195898749014),
-CLLocationCoordinate2D(latitude:47.7810834607949,longitude:-122.195830796072),
-CLLocationCoordinate2D(latitude:47.7810757268653,longitude:-122.195777894596),
-CLLocationCoordinate2D(latitude:47.7810757268653,longitude:-122.19572499312),
-CLLocationCoordinate2D(latitude:47.781070967524,longitude:-122.195667443396),
-CLLocationCoordinate2D(latitude:47.7810632335944,longitude:-122.195614320575),
-CLLocationCoordinate2D(latitude:47.7810562928884,longitude:-122.195560312374),
-CLLocationCoordinate2D(latitude:47.7810493521823,longitude:-122.195507410897),
-CLLocationCoordinate2D(latitude:47.7810392385821,longitude:-122.195455394801),
-CLLocationCoordinate2D(latitude:47.7810261503936,longitude:-122.195397845078),
-CLLocationCoordinate2D(latitude:47.7810152435698,longitude:-122.19533918863),
-CLLocationCoordinate2D(latitude:47.7810075096402,longitude:-122.195285180428),
-CLLocationCoordinate2D(latitude:47.7809966028164,longitude:-122.195232057607),
-CLLocationCoordinate2D(latitude:47.7809843078514,longitude:-122.195180262856),
-CLLocationCoordinate2D(latitude:47.7809702281335,longitude:-122.195129574831),
-CLLocationCoordinate2D(latitude:47.7809547602743,longitude:-122.195080214876),
-CLLocationCoordinate2D(latitude:47.780937705968,longitude:-122.195030633576),
-CLLocationCoordinate2D(latitude:47.780923824556,longitude:-122.194973083853),
-CLLocationCoordinate2D(latitude:47.7809091499203,longitude:-122.194922395828),
-CLLocationCoordinate2D(latitude:47.780890509167,longitude:-122.194876356049),
-CLLocationCoordinate2D(latitude:47.7808702819665,longitude:-122.194831422996),
-CLLocationCoordinate2D(latitude:47.7808524344367,longitude:-122.194785383218),
-CLLocationCoordinate2D(latitude:47.7808345869069,longitude:-122.194739343439),
-],
-[
-CLLocationCoordinate2D(latitude:47.782469618944,longitude:-122.196217707287),
-CLLocationCoordinate2D(latitude:47.7824829054384,longitude:-122.196168125987),
-CLLocationCoordinate2D(latitude:47.782490639368,longitude:-122.19611301106),
-CLLocationCoordinate2D(latitude:47.7824999597447,longitude:-122.196061216309),
-CLLocationCoordinate2D(latitude:47.7825114614861,longitude:-122.195999018339),
-CLLocationCoordinate2D(latitude:47.7825348615808,longitude:-122.195955192012),
-CLLocationCoordinate2D(latitude:47.782557270146,longitude:-122.19591380048),
-CLLocationCoordinate2D(latitude:47.7825828516054,longitude:-122.195843412742),
-CLLocationCoordinate2D(latitude:47.7825983194646,longitude:-122.195792946062),
-CLLocationCoordinate2D(latitude:47.7826108127354,longitude:-122.195743364762),
-CLLocationCoordinate2D(latitude:47.7826231077004,longitude:-122.195690463286),
-],
-[
-CLLocationCoordinate2D(latitude:47.783708832431,longitude:-122.196915387007),
-CLLocationCoordinate2D(latitude:47.7837110137957,longitude:-122.196861378805),
-CLLocationCoordinate2D(latitude:47.7837141866899,longitude:-122.196806042533),
-CLLocationCoordinate2D(latitude:47.783715773137,longitude:-122.196752034331),
-CLLocationCoordinate2D(latitude:47.783715773137,longitude:-122.196697804785),
-CLLocationCoordinate2D(latitude:47.7837312409962,longitude:-122.19664844483),
-CLLocationCoordinate2D(latitude:47.7837512698908,longitude:-122.196604618502),
-CLLocationCoordinate2D(latitude:47.7837784377973,longitude:-122.196542420532),
-CLLocationCoordinate2D(latitude:47.7837939056564,longitude:-122.196492839232),
-CLLocationCoordinate2D(latitude:47.7838095718215,longitude:-122.196443479278),
-CLLocationCoordinate2D(latitude:47.7838272210454,longitude:-122.196395004703),
-CLLocationCoordinate2D(latitude:47.7838442753517,longitude:-122.196347858199),
-CLLocationCoordinate2D(latitude:47.7838722364817,longitude:-122.196278798532),
-],
-]
-
