--- PhotoPoints/View/Map/Map Feature Data/Trails.swift
@@ -1,886 +1,0 @@
-//
-//  Trails.swift
-//  CoordinateDrawer
-//
-//  Created by Clay Suttner on 12/22/20.
-//
-
-import Foundation
-import MapKit
-
-class TrailLine: MKPolyline {}
-
-let trails: [[CLLocationCoordinate2D]] = [
-[
-CLLocationCoordinate2D(latitude:47.7742496416279,longitude:-122.191293000408),
-CLLocationCoordinate2D(latitude:47.7742472619573,longitude:-122.191346123229),
-CLLocationCoordinate2D(latitude:47.7742466670396,longitude:-122.191402566757),
-CLLocationCoordinate2D(latitude:47.7742466670396,longitude:-122.191473838814),
-CLLocationCoordinate2D(latitude:47.7742466670396,longitude:-122.191529176148),
-CLLocationCoordinate2D(latitude:47.7742504348514,longitude:-122.19158318435),
-CLLocationCoordinate2D(latitude:47.7742557891104,longitude:-122.191639626817),
-CLLocationCoordinate2D(latitude:47.7742629281223,longitude:-122.191706473033),
-CLLocationCoordinate2D(latitude:47.7742674891577,longitude:-122.191761809306),
-CLLocationCoordinate2D(latitude:47.7742736366402,longitude:-122.191818251772),
-CLLocationCoordinate2D(latitude:47.7742752230873,longitude:-122.191879122203),
-CLLocationCoordinate2D(latitude:47.7742799824286,longitude:-122.191943754969),
-CLLocationCoordinate2D(latitude:47.7742799824286,longitude:-122.192002410886),
-CLLocationCoordinate2D(latitude:47.7742799824286,longitude:-122.192056419088),
-CLLocationCoordinate2D(latitude:47.7742799824286,longitude:-122.1921369887),
-CLLocationCoordinate2D(latitude:47.7742813705698,longitude:-122.192192326034),
-CLLocationCoordinate2D(latitude:47.7742813705698,longitude:-122.192249875757),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.192305210968),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.192360548301),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.192415662698),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.19247099897),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.192546919805),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.192602257138),
-CLLocationCoordinate2D(latitude:47.7742821637933,longitude:-122.192655158084),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.192711600551),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.192774906308),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.192831348774),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.192885578852),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.192968360854),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.193023698187),
-CLLocationCoordinate2D(latitude:47.7742829570169,longitude:-122.193079033398),
-CLLocationCoordinate2D(latitude:47.7742861299111,longitude:-122.193135476926),
-CLLocationCoordinate2D(latitude:47.7742861299111,longitude:-122.193208963496),
-CLLocationCoordinate2D(latitude:47.7742805773462,longitude:-122.193268948546),
-CLLocationCoordinate2D(latitude:47.7742736366402,longitude:-122.193324283756),
-CLLocationCoordinate2D(latitude:47.7742643162635,longitude:-122.193377185763),
-CLLocationCoordinate2D(latitude:47.774251228075,longitude:-122.193440490459),
-CLLocationCoordinate2D(latitude:47.7742419076983,longitude:-122.193493391404),
-CLLocationCoordinate2D(latitude:47.7742310008745,longitude:-122.193545187216),
-CLLocationCoordinate2D(latitude:47.7742163262389,longitude:-122.193594768516),
-CLLocationCoordinate2D(latitude:47.7742016516033,longitude:-122.193646562205),
-CLLocationCoordinate2D(latitude:47.7741869769677,longitude:-122.193696144036),
-CLLocationCoordinate2D(latitude:47.7741627836495,longitude:-122.193737535568),
-CLLocationCoordinate2D(latitude:47.7741358140488,longitude:-122.193776713648),
-CLLocationCoordinate2D(latitude:47.7741070596953,longitude:-122.193811243482),
-CLLocationCoordinate2D(latitude:47.7740721578592,longitude:-122.193822753427),
-CLLocationCoordinate2D(latitude:47.7740396356937,longitude:-122.193846880042),
-CLLocationCoordinate2D(latitude:47.7740071135283,longitude:-122.193868793205),
-CLLocationCoordinate2D(latitude:47.7739706252451,longitude:-122.193905757835),
-CLLocationCoordinate2D(latitude:47.7739474234563,longitude:-122.193949362818),
-CLLocationCoordinate2D(latitude:47.7739258081147,longitude:-122.19399208242),
-CLLocationCoordinate2D(latitude:47.7738970537611,longitude:-122.194042770445),
-CLLocationCoordinate2D(latitude:47.7738784130077,longitude:-122.194087703498),
-CLLocationCoordinate2D(latitude:47.7738613587014,longitude:-122.194133743277),
-CLLocationCoordinate2D(latitude:47.7738669112663,longitude:-122.194179783055),
-CLLocationCoordinate2D(latitude:47.7739026063259,longitude:-122.194191293),
-CLLocationCoordinate2D(latitude:47.7739343352678,longitude:-122.194220067861),
-CLLocationCoordinate2D(latitude:47.7739660642097,longitude:-122.194244194476),
-CLLocationCoordinate2D(latitude:47.7739977931516,longitude:-122.194270755887),
-CLLocationCoordinate2D(latitude:47.7740404289173,longitude:-122.194307499171),
-CLLocationCoordinate2D(latitude:47.7740721578592,longitude:-122.194334060582),
-CLLocationCoordinate2D(latitude:47.7741040851069,longitude:-122.194361728718),
-CLLocationCoordinate2D(latitude:47.7741334343782,longitude:-122.19439271703),
-CLLocationCoordinate2D(latitude:47.7741822176264,longitude:-122.194425033413),
-CLLocationCoordinate2D(latitude:47.7742131533447,longitude:-122.194452701549),
-CLLocationCoordinate2D(latitude:47.7742450805925,longitude:-122.19447793489),
-CLLocationCoordinate2D(latitude:47.7742783959815,longitude:-122.194496306532),
-CLLocationCoordinate2D(latitude:47.7743101249234,longitude:-122.194498741328),
-CLLocationCoordinate2D(latitude:47.774324799559,longitude:-122.194450266753),
-CLLocationCoordinate2D(latitude:47.774345819983,longitude:-122.194406440426),
-CLLocationCoordinate2D(latitude:47.7743472081242,longitude:-122.19435353895),
-CLLocationCoordinate2D(latitude:47.7743394741946,longitude:-122.194279831035),
-CLLocationCoordinate2D(latitude:47.7743357063828,longitude:-122.194226929559),
-CLLocationCoordinate2D(latitude:47.7743349131592,longitude:-122.194174028083),
-CLLocationCoordinate2D(latitude:47.7743333267121,longitude:-122.194119798536),
-CLLocationCoordinate2D(latitude:47.7743317402651,longitude:-122.194056493841),
-CLLocationCoordinate2D(latitude:47.7743317402651,longitude:-122.194002485639),
-CLLocationCoordinate2D(latitude:47.7743317402651,longitude:-122.193949362818),
-CLLocationCoordinate2D(latitude:47.7743357063828,longitude:-122.193894247891),
-CLLocationCoordinate2D(latitude:47.7743325334886,longitude:-122.193826294948),
-CLLocationCoordinate2D(latitude:47.7743039774409,longitude:-122.193845773316),
-CLLocationCoordinate2D(latitude:47.7742813705698,longitude:-122.193886058122),
-CLLocationCoordinate2D(latitude:47.7742480551808,longitude:-122.193903323039),
-CLLocationCoordinate2D(latitude:47.7742139465683,longitude:-122.193884951397),
-CLLocationCoordinate2D(latitude:47.7741838040735,longitude:-122.193859718057),
-CLLocationCoordinate2D(latitude:47.7741544548022,longitude:-122.193829615125),
-],
-[
-CLLocationCoordinate2D(latitude:47.7743930167841,longitude:-122.196309786844),
-CLLocationCoordinate2D(latitude:47.7743860760781,longitude:-122.196256885368),
-CLLocationCoordinate2D(latitude:47.774379135372,longitude:-122.196202655821),
-CLLocationCoordinate2D(latitude:47.7743682285482,longitude:-122.196132489428),
-CLLocationCoordinate2D(latitude:47.7743612878422,longitude:-122.196078481226),
-CLLocationCoordinate2D(latitude:47.7743495877949,longitude:-122.196026686476),
-CLLocationCoordinate2D(latitude:47.774338879277,longitude:-122.19597599845),
-CLLocationCoordinate2D(latitude:47.7743170656294,longitude:-122.195910258959),
-CLLocationCoordinate2D(latitude:47.7743085384763,longitude:-122.195860677659),
-CLLocationCoordinate2D(latitude:47.7743287656768,longitude:-122.19581463788),
-CLLocationCoordinate2D(latitude:47.7743480013478,longitude:-122.195767491377),
-CLLocationCoordinate2D(latitude:47.7743706082189,longitude:-122.195697324983),
-CLLocationCoordinate2D(latitude:47.7743813167368,longitude:-122.195646636958),
-CLLocationCoordinate2D(latitude:47.7743975778195,longitude:-122.195597055658),
-CLLocationCoordinate2D(latitude:47.7744116575375,longitude:-122.195546367633),
-CLLocationCoordinate2D(latitude:47.7744231592789,longitude:-122.195494572882),
-CLLocationCoordinate2D(latitude:47.7744372389969,longitude:-122.195443884857),
-CLLocationCoordinate2D(latitude:47.77444576615,longitude:-122.195392090106),
-CLLocationCoordinate2D(latitude:47.7744511204089,longitude:-122.195323030438),
-CLLocationCoordinate2D(latitude:47.774458061115,longitude:-122.195270128962),
-CLLocationCoordinate2D(latitude:47.7744628204563,longitude:-122.195217227486),
-CLLocationCoordinate2D(latitude:47.7744628204563,longitude:-122.195157242967),
-CLLocationCoordinate2D(latitude:47.7744550865267,longitude:-122.195104341491),
-CLLocationCoordinate2D(latitude:47.7744418000323,longitude:-122.195053653465),
-CLLocationCoordinate2D(latitude:47.774431686432,longitude:-122.195000751989),
-CLLocationCoordinate2D(latitude:47.7744295050673,longitude:-122.1949281508),
-CLLocationCoordinate2D(latitude:47.7744310915144,longitude:-122.194875249324),
-CLLocationCoordinate2D(latitude:47.7744364457733,longitude:-122.194821019777),
-CLLocationCoordinate2D(latitude:47.7744394203616,longitude:-122.194768118301),
-CLLocationCoordinate2D(latitude:47.7744394203616,longitude:-122.194715216825),
-CLLocationCoordinate2D(latitude:47.7744287118437,longitude:-122.194664528799),
-CLLocationCoordinate2D(latitude:47.7744047168314,longitude:-122.194622915923),
-CLLocationCoordinate2D(latitude:47.7743658488776,longitude:-122.194571121172),
-CLLocationCoordinate2D(latitude:47.774312504594,longitude:-122.194522867943),
-],
-[
-CLLocationCoordinate2D(latitude:47.7743799285956,longitude:-122.19576284313),
-CLLocationCoordinate2D(latitude:47.7744162185729,longitude:-122.195768598102),
-CLLocationCoordinate2D(latitude:47.7744519136325,longitude:-122.195779001321),
-CLLocationCoordinate2D(latitude:47.7744852290215,longitude:-122.195797372964),
-CLLocationCoordinate2D(latitude:47.7745153715163,longitude:-122.1958250411),
-CLLocationCoordinate2D(latitude:47.7745433326463,longitude:-122.195859570934),
-CLLocationCoordinate2D(latitude:47.7745651462939,longitude:-122.195918448727),
-CLLocationCoordinate2D(latitude:47.774558007282,longitude:-122.195970243478),
-CLLocationCoordinate2D(latitude:47.7745494801288,longitude:-122.196022038229),
-CLLocationCoordinate2D(latitude:47.7745417461992,longitude:-122.196074939705),
-CLLocationCoordinate2D(latitude:47.7745518597995,longitude:-122.196134924224),
-CLLocationCoordinate2D(latitude:47.774579027706,longitude:-122.196169454058),
-CLLocationCoordinate2D(latitude:47.774606988836,longitude:-122.196202655821),
-CLLocationCoordinate2D(latitude:47.7746339584366,longitude:-122.196237407),
-CLLocationCoordinate2D(latitude:47.7746627127902,longitude:-122.196276363736),
-CLLocationCoordinate2D(latitude:47.7746912688379,longitude:-122.196308680119),
-CLLocationCoordinate2D(latitude:47.7747168502973,longitude:-122.196278798532),
-CLLocationCoordinate2D(latitude:47.774747191098,longitude:-122.196251130395),
-CLLocationCoordinate2D(latitude:47.7747828861577,longitude:-122.196253343846),
-CLLocationCoordinate2D(latitude:47.7748162015466,longitude:-122.196274150285),
-CLLocationCoordinate2D(latitude:47.7748501118533,longitude:-122.19628787368),
-CLLocationCoordinate2D(latitude:47.7748943340661,longitude:-122.19628787368),
-CLLocationCoordinate2D(latitude:47.7749268562315,longitude:-122.196269502038),
-CLLocationCoordinate2D(latitude:47.7749617580676,longitude:-122.196255778642),
-CLLocationCoordinate2D(latitude:47.7750091531746,longitude:-122.196263747066),
-CLLocationCoordinate2D(latitude:47.7750454431518,longitude:-122.196266181862),
-CLLocationCoordinate2D(latitude:47.7750827246586,longitude:-122.196266181862),
-CLLocationCoordinate2D(latitude:47.7751176264947,longitude:-122.196255778642),
-CLLocationCoordinate2D(latitude:47.7751555029191,longitude:-122.196229217232),
-CLLocationCoordinate2D(latitude:47.7751834640491,longitude:-122.196194687398),
-CLLocationCoordinate2D(latitude:47.7752183658852,longitude:-122.196188932426),
-CLLocationCoordinate2D(latitude:47.775243154121,longitude:-122.196228110506),
-CLLocationCoordinate2D(latitude:47.7752687355805,longitude:-122.196266181862),
-CLLocationCoordinate2D(latitude:47.7752988780753,longitude:-122.196297170174),
-CLLocationCoordinate2D(latitude:47.7753345731349,longitude:-122.196302925146),
-CLLocationCoordinate2D(latitude:47.7753700698886,longitude:-122.196302925146),
-CLLocationCoordinate2D(latitude:47.7754065581718,longitude:-122.196302925146),
-CLLocationCoordinate2D(latitude:47.7754390803373,longitude:-122.196336348255),
-CLLocationCoordinate2D(latitude:47.7754702143615,longitude:-122.196360474869),
-CLLocationCoordinate2D(latitude:47.7755049178917,longitude:-122.19637441961),
-CLLocationCoordinate2D(latitude:47.7755467604338,longitude:-122.196382388033),
-CLLocationCoordinate2D(latitude:47.7755824554935,longitude:-122.196377739786),
-CLLocationCoordinate2D(latitude:47.7756181505531,longitude:-122.196383494759),
-CLLocationCoordinate2D(latitude:47.7756584066481,longitude:-122.19641226962),
-CLLocationCoordinate2D(latitude:47.7756863677782,longitude:-122.196450340975),
-CLLocationCoordinate2D(latitude:47.775711156014,longitude:-122.196490625782),
-CLLocationCoordinate2D(latitude:47.7757359442499,longitude:-122.196529803862),
-CLLocationCoordinate2D(latitude:47.7757637070741,longitude:-122.196596650079),
-CLLocationCoordinate2D(latitude:47.7757886936158,longitude:-122.196634500089),
-CLLocationCoordinate2D(latitude:47.7758212157812,longitude:-122.196656413253),
-CLLocationCoordinate2D(latitude:47.775856712535,longitude:-122.196665709747),
-CLLocationCoordinate2D(latitude:47.7758947872653,longitude:-122.196695591334),
-CLLocationCoordinate2D(latitude:47.7759142212422,longitude:-122.196740524387),
-CLLocationCoordinate2D(latitude:47.7759281026543,longitude:-122.196789884342),
-CLLocationCoordinate2D(latitude:47.775939009478,longitude:-122.196844113888),
-CLLocationCoordinate2D(latitude:47.7759396043957,longitude:-122.196916715077),
-],
-[
-CLLocationCoordinate2D(latitude:47.7759078754538,longitude:-122.196655306528),
-CLLocationCoordinate2D(latitude:47.7759411908428,longitude:-122.196634500089),
-CLLocationCoordinate2D(latitude:47.7759707384199,longitude:-122.196606831953),
-CLLocationCoordinate2D(latitude:47.7759893791733,longitude:-122.196545962054),
-CLLocationCoordinate2D(latitude:47.7759923537616,longitude:-122.196492839232),
-CLLocationCoordinate2D(latitude:47.7759955266558,longitude:-122.196439937756),
-CLLocationCoordinate2D(latitude:47.7759971131029,longitude:-122.196367336567),
-CLLocationCoordinate2D(latitude:47.7760155555504,longitude:-122.196321296789),
-CLLocationCoordinate2D(latitude:47.7760373691979,longitude:-122.196279905257),
-CLLocationCoordinate2D(latitude:47.7760403437862,longitude:-122.196227003781),
-CLLocationCoordinate2D(latitude:47.7760234877858,longitude:-122.196177422481),
-CLLocationCoordinate2D(latitude:47.7760312217154,longitude:-122.196124521005),
-CLLocationCoordinate2D(latitude:47.776010994515,longitude:-122.196082908128),
-CLLocationCoordinate2D(latitude:47.7759784723495,longitude:-122.196058781513),
-CLLocationCoordinate2D(latitude:47.7759606248197,longitude:-122.195989721846),
-CLLocationCoordinate2D(latitude:47.775956658702,longitude:-122.195936820369),
-CLLocationCoordinate2D(latitude:47.7759326636896,longitude:-122.195898749014),
-CLLocationCoordinate2D(latitude:47.7759009347478,longitude:-122.195872408949),
-CLLocationCoordinate2D(latitude:47.7758707922529,longitude:-122.195843412742),
-CLLocationCoordinate2D(latitude:47.7758622650998,longitude:-122.195796266238),
-CLLocationCoordinate2D(latitude:47.7758815007708,longitude:-122.195751333185),
-CLLocationCoordinate2D(latitude:47.7758947872653,longitude:-122.195703079956),
-CLLocationCoordinate2D(latitude:47.7759017279713,longitude:-122.195637340464),
-CLLocationCoordinate2D(latitude:47.7759041076419,longitude:-122.195584438988),
-CLLocationCoordinate2D(latitude:47.7759049008655,longitude:-122.195531537512),
-CLLocationCoordinate2D(latitude:47.7758862601121,longitude:-122.195480849487),
-CLLocationCoordinate2D(latitude:47.7758707922529,longitude:-122.195431268187),
-CLLocationCoordinate2D(latitude:47.7758467972406,longitude:-122.195392090106),
-CLLocationCoordinate2D(latitude:47.7758249835931,longitude:-122.1953518053),
-CLLocationCoordinate2D(latitude:47.7758172496635,longitude:-122.195295362302),
-CLLocationCoordinate2D(latitude:47.7758055496162,longitude:-122.195237812579),
-CLLocationCoordinate2D(latitude:47.7757908749805,longitude:-122.195188452624),
-CLLocationCoordinate2D(latitude:47.7757793732391,longitude:-122.195137764599),
-CLLocationCoordinate2D(latitude:47.7757692596389,longitude:-122.195083535052),
-CLLocationCoordinate2D(latitude:47.775756766368,longitude:-122.195032847027),
-CLLocationCoordinate2D(latitude:47.7757583528151,longitude:-122.194979945551),
-CLLocationCoordinate2D(latitude:47.7757785800156,longitude:-122.194936119223),
-CLLocationCoordinate2D(latitude:47.7757684664153,longitude:-122.194899375939),
-CLLocationCoordinate2D(latitude:47.7757436781795,longitude:-122.194875249324),
-CLLocationCoordinate2D(latitude:47.775744471403,longitude:-122.194822126503),
-CLLocationCoordinate2D(latitude:47.7757242442026,longitude:-122.194756608356),
-CLLocationCoordinate2D(latitude:47.7757010424138,longitude:-122.19471632355),
-CLLocationCoordinate2D(latitude:47.7756677270248,longitude:-122.194697951908),
-CLLocationCoordinate2D(latitude:47.77563758453,longitude:-122.194665635525),
-CLLocationCoordinate2D(latitude:47.7755949487643,longitude:-122.194647263882),
-CLLocationCoordinate2D(latitude:47.7755584604812,longitude:-122.194647263882),
-CLLocationCoordinate2D(latitude:47.7755211789744,longitude:-122.194647263882),
-CLLocationCoordinate2D(latitude:47.7755382332807,longitude:-122.194581524391),
-CLLocationCoordinate2D(latitude:47.7755507265516,longitude:-122.194530836366),
-CLLocationCoordinate2D(latitude:47.7755624265989,longitude:-122.194473286643),
-],
-[
-CLLocationCoordinate2D(latitude:47.7760078216208,longitude:-122.196488412331),
-CLLocationCoordinate2D(latitude:47.7760341963037,longitude:-122.196451447701),
-CLLocationCoordinate2D(latitude:47.7760637438809,longitude:-122.196422672839),
-CLLocationCoordinate2D(latitude:47.7760970592699,longitude:-122.19639965295),
-CLLocationCoordinate2D(latitude:47.7761357289178,longitude:-122.196391684527),
-CLLocationCoordinate2D(latitude:47.776158137483,longitude:-122.196431969333),
-CLLocationCoordinate2D(latitude:47.7761738036481,longitude:-122.196480222562),
-CLLocationCoordinate2D(latitude:47.7761876850601,longitude:-122.196530910588),
-CLLocationCoordinate2D(latitude:47.7761962122133,longitude:-122.196582705338),
-CLLocationCoordinate2D(latitude:47.7761970054368,longitude:-122.196635606815),
-CLLocationCoordinate2D(latitude:47.7762225868962,longitude:-122.196672571444),
-CLLocationCoordinate2D(latitude:47.7762233801198,longitude:-122.196725472921),
-CLLocationCoordinate2D(latitude:47.776219414002,longitude:-122.196777267671),
-CLLocationCoordinate2D(latitude:47.7761892715072,longitude:-122.196806042533),
-CLLocationCoordinate2D(latitude:47.7761496103299,longitude:-122.196801394286),
-CLLocationCoordinate2D(latitude:47.7761141135761,longitude:-122.196794532588),
-CLLocationCoordinate2D(latitude:47.77607921174,longitude:-122.196776160946),
-CLLocationCoordinate2D(latitude:47.7760443099039,longitude:-122.196760002755),
-CLLocationCoordinate2D(latitude:47.7759979063264,longitude:-122.196745172634),
-CLLocationCoordinate2D(latitude:47.7759528908901,longitude:-122.196770405974),
-],
-[
-CLLocationCoordinate2D(latitude:47.7765781493514,longitude:-122.195828582621),
-CLLocationCoordinate2D(latitude:47.7766130511875,longitude:-122.19581463788),
-CLLocationCoordinate2D(latitude:47.7766370461998,longitude:-122.195775681145),
-CLLocationCoordinate2D(latitude:47.7766572734002,longitude:-122.195731854817),
-CLLocationCoordinate2D(latitude:47.77667512093,longitude:-122.195681166792),
-CLLocationCoordinate2D(latitude:47.7766921752363,longitude:-122.195635127014),
-CLLocationCoordinate2D(latitude:47.7767044702013,longitude:-122.195585545714),
-CLLocationCoordinate2D(latitude:47.7767324313313,longitude:-122.195539505935),
-CLLocationCoordinate2D(latitude:47.7767689196145,longitude:-122.195532644237),
-CLLocationCoordinate2D(latitude:47.7768044163683,longitude:-122.195532644237),
-CLLocationCoordinate2D(latitude:47.7768409046515,longitude:-122.195532644237),
-CLLocationCoordinate2D(latitude:47.7768758064875,longitude:-122.195521134293),
-CLLocationCoordinate2D(latitude:47.7769021811705,longitude:-122.195486604459),
-CLLocationCoordinate2D(latitude:47.7769184422532,longitude:-122.195438129885),
-CLLocationCoordinate2D(latitude:47.7769370830066,longitude:-122.195393418177),
-CLLocationCoordinate2D(latitude:47.7769517576422,longitude:-122.195343836877),
-CLLocationCoordinate2D(latitude:47.7769725797603,longitude:-122.195302445345),
-CLLocationCoordinate2D(latitude:47.7770090680435,longitude:-122.195301117275),
-CLLocationCoordinate2D(latitude:47.7770461512444,longitude:-122.195303552071),
-CLLocationCoordinate2D(latitude:47.7770965209396,longitude:-122.195310413768),
-CLLocationCoordinate2D(latitude:47.7771330092228,longitude:-122.195312627219),
-CLLocationCoordinate2D(latitude:47.7771687042824,longitude:-122.19531395529),
-CLLocationCoordinate2D(latitude:47.777202812895,longitude:-122.195329892136),
-CLLocationCoordinate2D(latitude:47.7772321621662,longitude:-122.195358666998),
-CLLocationCoordinate2D(latitude:47.7772561571785,longitude:-122.195395631628),
-CLLocationCoordinate2D(latitude:47.7772918522382,longitude:-122.195400279874),
-CLLocationCoordinate2D(latitude:47.7773221930389,longitude:-122.195373718464),
-CLLocationCoordinate2D(latitude:47.7773313151097,longitude:-122.195323030438),
-CLLocationCoordinate2D(latitude:47.7773616559103,longitude:-122.195294255577),
-CLLocationCoordinate2D(latitude:47.77739735097,longitude:-122.195294255577),
-CLLocationCoordinate2D(latitude:47.7774437545475,longitude:-122.195305765521),
-CLLocationCoordinate2D(latitude:47.77747786316,longitude:-122.195325465234),
-CLLocationCoordinate2D(latitude:47.777511178549,longitude:-122.195344943602),
-CLLocationCoordinate2D(latitude:47.7775351735613,longitude:-122.195384121683),
-CLLocationCoordinate2D(latitude:47.7775437007145,longitude:-122.195437023159),
-CLLocationCoordinate2D(latitude:47.7775746364328,longitude:-122.19546358457),
-CLLocationCoordinate2D(latitude:47.7776087450454,longitude:-122.195477307965),
-CLLocationCoordinate2D(latitude:47.777651380811,longitude:-122.195498114404),
-CLLocationCoordinate2D(latitude:47.7776817216117,longitude:-122.19552578254),
-CLLocationCoordinate2D(latitude:47.7777009572827,longitude:-122.195570715593),
-CLLocationCoordinate2D(latitude:47.7777065098476,longitude:-122.19562583052),
-CLLocationCoordinate2D(latitude:47.7777142437772,longitude:-122.195684708313),
-CLLocationCoordinate2D(latitude:47.7777257455186,longitude:-122.195735174994),
-CLLocationCoordinate2D(latitude:47.7777499388368,longitude:-122.195773246349),
-CLLocationCoordinate2D(latitude:47.7777894017083,longitude:-122.19579980776),
-CLLocationCoordinate2D(latitude:47.7778219238737,longitude:-122.195822827649),
-CLLocationCoordinate2D(latitude:47.7778576189334,longitude:-122.195835444319),
-CLLocationCoordinate2D(latitude:47.7778931156871,longitude:-122.195850495785),
-CLLocationCoordinate2D(latitude:47.777933570088,longitude:-122.1958746224),
-CLLocationCoordinate2D(latitude:47.7779652990299,longitude:-122.195897642289),
-CLLocationCoordinate2D(latitude:47.7779954415247,longitude:-122.195929958672),
-CLLocationCoordinate2D(latitude:47.7780257823254,longitude:-122.195960946984),
-CLLocationCoordinate2D(latitude:47.7780604858556,longitude:-122.195983966873),
-CLLocationCoordinate2D(latitude:47.7780969741388,longitude:-122.195996583543),
-CLLocationCoordinate2D(latitude:47.7781318759749,longitude:-122.196011635009),
-CLLocationCoordinate2D(latitude:47.7781636049168,longitude:-122.196040409871),
-CLLocationCoordinate2D(latitude:47.7781822456702,longitude:-122.196085342924),
-CLLocationCoordinate2D(latitude:47.7782046542354,longitude:-122.196130275977),
-CLLocationCoordinate2D(latitude:47.7782363831773,longitude:-122.196155509317),
-CLLocationCoordinate2D(latitude:47.7782635510838,longitude:-122.196188932426),
-CLLocationCoordinate2D(latitude:47.7782930986609,longitude:-122.196218814013),
-CLLocationCoordinate2D(latitude:47.7783272072734,longitude:-122.196240727176),
-CLLocationCoordinate2D(latitude:47.778361315886,longitude:-122.196257992093),
-CLLocationCoordinate2D(latitude:47.7783983990868,longitude:-122.196266181862),
-CLLocationCoordinate2D(latitude:47.77843488737,longitude:-122.196266181862),
-CLLocationCoordinate2D(latitude:47.7784705824296,longitude:-122.196257992093),
-CLLocationCoordinate2D(latitude:47.7785052859598,longitude:-122.196240727176),
-CLLocationCoordinate2D(latitude:47.7785463352784,longitude:-122.196222355534),
-CLLocationCoordinate2D(latitude:47.7785820303381,longitude:-122.196215493836),
-CLLocationCoordinate2D(latitude:47.7786185186212,longitude:-122.196215493836),
-CLLocationCoordinate2D(latitude:47.7786635340576,longitude:-122.196215493836),
-CLLocationCoordinate2D(latitude:47.7786990308113,longitude:-122.196221248809),
-CLLocationCoordinate2D(latitude:47.7787339326474,longitude:-122.196234972204),
-CLLocationCoordinate2D(latitude:47.7787704209306,longitude:-122.196240727176),
-CLLocationCoordinate2D(latitude:47.7788075041314,longitude:-122.196231652028),
-CLLocationCoordinate2D(latitude:47.7788346720379,longitude:-122.196198228919),
-CLLocationCoordinate2D(latitude:47.7788648145327,longitude:-122.196168125987),
-CLLocationCoordinate2D(latitude:47.7788935688863,longitude:-122.196137137675),
-CLLocationCoordinate2D(latitude:47.7789191503457,longitude:-122.196101501116),
-CLLocationCoordinate2D(latitude:47.7789524657347,longitude:-122.196081801403),
-CLLocationCoordinate2D(latitude:47.7789881607943,longitude:-122.196081801403),
-CLLocationCoordinate2D(latitude:47.779023855854,longitude:-122.196072726254),
-CLLocationCoordinate2D(latitude:47.7790407118544,longitude:-122.19602425168),
-CLLocationCoordinate2D(latitude:47.7790633187255,longitude:-122.195981753422),
-CLLocationCoordinate2D(latitude:47.7790942544438,longitude:-122.195952978561),
-CLLocationCoordinate2D(latitude:47.7791329240917,longitude:-122.195945895518),
-CLLocationCoordinate2D(latitude:47.7791702055985,longitude:-122.195948330314),
-CLLocationCoordinate2D(latitude:47.7792051074346,longitude:-122.195958733533),
-CLLocationCoordinate2D(latitude:47.7792392160471,longitude:-122.195979318627),
-CLLocationCoordinate2D(latitude:47.7792554771298,longitude:-122.19602425168),
-CLLocationCoordinate2D(latitude:47.7792701517654,longitude:-122.196054133267),
-CLLocationCoordinate2D(latitude:47.7793050536015,longitude:-122.19603819642),
-CLLocationCoordinate2D(latitude:47.7793562165203,longitude:-122.196011635009),
-CLLocationCoordinate2D(latitude:47.77939191158,longitude:-122.196006986763),
-CLLocationCoordinate2D(latitude:47.7794282015573,longitude:-122.196006986763),
-CLLocationCoordinate2D(latitude:47.7794638966169,longitude:-122.196012741735),
-CLLocationCoordinate2D(latitude:47.7794995916765,longitude:-122.19597599845),
-CLLocationCoordinate2D(latitude:47.7795103001944,longitude:-122.195925310425),
-CLLocationCoordinate2D(latitude:47.7795289409478,longitude:-122.195878163921),
-CLLocationCoordinate2D(latitude:47.7795475817012,longitude:-122.195819286127),
-CLLocationCoordinate2D(latitude:47.7795739563841,longitude:-122.195784756294),
-],
-[
-CLLocationCoordinate2D(latitude:47.7749579902557,longitude:-122.193412821792),
-CLLocationCoordinate2D(latitude:47.7749625512911,longitude:-122.193359920846),
-CLLocationCoordinate2D(latitude:47.774975044562,longitude:-122.19330923229),
-CLLocationCoordinate2D(latitude:47.7749881327506,longitude:-122.193258544796),
-CLLocationCoordinate2D(latitude:47.7750113345393,longitude:-122.193208963496),
-CLLocationCoordinate2D(latitude:47.7750353295516,longitude:-122.19316867869),
-CLLocationCoordinate2D(latitude:47.7750563499756,longitude:-122.193125073176),
-CLLocationCoordinate2D(latitude:47.7750787585408,longitude:-122.1930834603),
-CLLocationCoordinate2D(latitude:47.7750989857413,longitude:-122.193035208132),
-CLLocationCoordinate2D(latitude:47.7751182214123,longitude:-122.192989168353),
-CLLocationCoordinate2D(latitude:47.7751424147305,longitude:-122.192949989742),
-CLLocationCoordinate2D(latitude:47.7751709707782,longitude:-122.19290505722),
-CLLocationCoordinate2D(latitude:47.7751935776493,longitude:-122.192863665157),
-CLLocationCoordinate2D(latitude:47.7752112268733,longitude:-122.192817625379),
-CLLocationCoordinate2D(latitude:47.7752253065912,longitude:-122.192768044079),
-CLLocationCoordinate2D(latitude:47.7752354201915,longitude:-122.192709388161),
-CLLocationCoordinate2D(latitude:47.7752407744504,longitude:-122.192655158084),
-CLLocationCoordinate2D(latitude:47.7752453354858,longitude:-122.192603363333),
-CLLocationCoordinate2D(latitude:47.7752477151564,longitude:-122.192549355131),
-CLLocationCoordinate2D(latitude:47.7752453354858,longitude:-122.192473212421),
-CLLocationCoordinate2D(latitude:47.7752245133677,longitude:-122.192429607969),
-CLLocationCoordinate2D(latitude:47.7752013115789,longitude:-122.192388215907),
-CLLocationCoordinate2D(latitude:47.7751648232957,longitude:-122.192332879635),
-CLLocationCoordinate2D(latitude:47.7751368621657,longitude:-122.192299457057),
-CLLocationCoordinate2D(latitude:47.7751198078594,longitude:-122.192252310022),
-CLLocationCoordinate2D(latitude:47.7751005721884,longitude:-122.1922073775),
-CLLocationCoordinate2D(latitude:47.7750849060233,longitude:-122.192158902395),
-CLLocationCoordinate2D(latitude:47.7750686449406,longitude:-122.192110649165),
-CLLocationCoordinate2D(latitude:47.7750555567521,longitude:-122.192061067866),
-CLLocationCoordinate2D(latitude:47.7750424685636,longitude:-122.192010379309),
-CLLocationCoordinate2D(latitude:47.7750369159987,longitude:-122.191956371108),
-CLLocationCoordinate2D(latitude:47.7750361227752,longitude:-122.191903248287),
-CLLocationCoordinate2D(latitude:47.7750361227752,longitude:-122.191850347341),
-CLLocationCoordinate2D(latitude:47.77504167534,longitude:-122.191785936451),
-CLLocationCoordinate2D(latitude:47.7750502024931,longitude:-122.191734140639),
-CLLocationCoordinate2D(latitude:47.775060911011,longitude:-122.19168234695),
-CLLocationCoordinate2D(latitude:47.7750763788702,longitude:-122.19163276565),
-CLLocationCoordinate2D(latitude:47.775108306118,longitude:-122.191552196037),
-CLLocationCoordinate2D(latitude:47.7751221875301,longitude:-122.191503720932),
-CLLocationCoordinate2D(latitude:47.7751213943065,longitude:-122.191450818925),
-CLLocationCoordinate2D(latitude:47.775087285694,longitude:-122.191437096591),
-CLLocationCoordinate2D(latitude:47.7750515906343,longitude:-122.191437096591),
-CLLocationCoordinate2D(latitude:47.775015300657,longitude:-122.191435989335),
-CLLocationCoordinate2D(latitude:47.7749796055974,longitude:-122.19143930898),
-CLLocationCoordinate2D(latitude:47.7749462902084,longitude:-122.19146232887),
-CLLocationCoordinate2D(latitude:47.7749339952434,longitude:-122.191509475904),
-CLLocationCoordinate2D(latitude:47.7749353833846,longitude:-122.191564812177),
-CLLocationCoordinate2D(latitude:47.7749316155728,longitude:-122.191620148449),
-CLLocationCoordinate2D(latitude:47.7749238816432,longitude:-122.191673049394),
-CLLocationCoordinate2D(latitude:47.7749028612192,longitude:-122.191716876783),
-CLLocationCoordinate2D(latitude:47.7748743051715,longitude:-122.19174897129),
-CLLocationCoordinate2D(latitude:47.7748106489818,longitude:-122.191790584167),
-CLLocationCoordinate2D(latitude:47.7747757471457,longitude:-122.191800987917),
-CLLocationCoordinate2D(latitude:47.774740250392,longitude:-122.191800987917),
-CLLocationCoordinate2D(latitude:47.7747091163677,longitude:-122.191776638896),
-CLLocationCoordinate2D(latitude:47.7746944417321,longitude:-122.191721524499),
-CLLocationCoordinate2D(latitude:47.7746992010734,longitude:-122.191668401678),
-CLLocationCoordinate2D(latitude:47.7747130824855,longitude:-122.191619041193),
-CLLocationCoordinate2D(latitude:47.7747331113801,longitude:-122.191571674405),
-CLLocationCoordinate2D(latitude:47.7747610725101,longitude:-122.191533823864),
-CLLocationCoordinate2D(latitude:47.774792801452,longitude:-122.191502614737),
-CLLocationCoordinate2D(latitude:47.7748269100645,longitude:-122.191480701043),
-CLLocationCoordinate2D(latitude:47.7748641915713,longitude:-122.191473838814),
-CLLocationCoordinate2D(latitude:47.7749105951488,longitude:-122.191484242564),
-],
-[
-CLLocationCoordinate2D(latitude:47.7750989857413,longitude:-122.192252310022),
-CLLocationCoordinate2D(latitude:47.775074990729,longitude:-122.192291488634),
-CLLocationCoordinate2D(latitude:47.7750515906343,longitude:-122.192330445369),
-CLLocationCoordinate2D(latitude:47.7750254142573,longitude:-122.192330445369),
-CLLocationCoordinate2D(latitude:47.7750012209391,longitude:-122.192289053307),
-CLLocationCoordinate2D(latitude:47.7749742513385,longitude:-122.192254523473),
-CLLocationCoordinate2D(latitude:47.7749423240907,longitude:-122.192229290133),
-CLLocationCoordinate2D(latitude:47.7749098019252,longitude:-122.1922073775),
-CLLocationCoordinate2D(latitude:47.7748756933127,longitude:-122.192192326034),
-CLLocationCoordinate2D(latitude:47.7748409897825,longitude:-122.192181922284),
-CLLocationCoordinate2D(latitude:47.7748223490291,longitude:-122.192138317832),
-CLLocationCoordinate2D(latitude:47.7748052947229,longitude:-122.192079438977),
-CLLocationCoordinate2D(latitude:47.7747834810753,longitude:-122.192035834525),
-CLLocationCoordinate2D(latitude:47.7747578996159,longitude:-122.191996656976),
-CLLocationCoordinate2D(latitude:47.7747309300153,longitude:-122.191959691815),
-CLLocationCoordinate2D(latitude:47.7747224028622,longitude:-122.191910331329),
-CLLocationCoordinate2D(latitude:47.7747301367917,longitude:-122.191826220196),
-],
-[
-CLLocationCoordinate2D(latitude:47.7763233262868,longitude:-122.191804307563),
-CLLocationCoordinate2D(latitude:47.7763596162641,longitude:-122.191818251772),
-CLLocationCoordinate2D(latitude:47.7763953113237,longitude:-122.191819359029),
-CLLocationCoordinate2D(latitude:47.7764302131598,longitude:-122.191803200306),
-CLLocationCoordinate2D(latitude:47.7764635285488,longitude:-122.19178239493),
-CLLocationCoordinate2D(latitude:47.7764992236084,longitude:-122.191784829195),
-CLLocationCoordinate2D(latitude:47.7765301593267,longitude:-122.191816924764),
-CLLocationCoordinate2D(latitude:47.7765603018216,longitude:-122.191857208508),
-CLLocationCoordinate2D(latitude:47.7765882629516,longitude:-122.191893067474),
-CLLocationCoordinate2D(latitude:47.7766130511875,longitude:-122.191932245024),
-CLLocationCoordinate2D(latitude:47.776640219094,longitude:-122.191966774858),
-CLLocationCoordinate2D(latitude:47.776668180224,longitude:-122.192004625399),
-CLLocationCoordinate2D(latitude:47.7766967362717,longitude:-122.192039155233),
-CLLocationCoordinate2D(latitude:47.7767246974018,longitude:-122.19207257781),
-CLLocationCoordinate2D(latitude:47.7767526585318,longitude:-122.192109541909),
-CLLocationCoordinate2D(latitude:47.7767774467677,longitude:-122.192148720521),
-CLLocationCoordinate2D(latitude:47.7768036231447,longitude:-122.19218657),
-CLLocationCoordinate2D(latitude:47.7768363436161,longitude:-122.192209589889),
-CLLocationCoordinate2D(latitude:47.7768702539227,longitude:-122.192225748612),
-CLLocationCoordinate2D(latitude:47.776915269359,longitude:-122.192231503584),
-CLLocationCoordinate2D(latitude:47.7769509644187,longitude:-122.1922361513),
-CLLocationCoordinate2D(latitude:47.7769811069135,longitude:-122.192264927223),
-CLLocationCoordinate2D(latitude:47.77700827482,longitude:-122.192298349801),
-CLLocationCoordinate2D(latitude:47.7770469444679,longitude:-122.192335093085),
-CLLocationCoordinate2D(latitude:47.7770810530804,longitude:-122.192358112975),
-CLLocationCoordinate2D(latitude:47.777115161693,longitude:-122.192375377892),
-CLLocationCoordinate2D(latitude:47.7771671178353,longitude:-122.192393970879),
-CLLocationCoordinate2D(latitude:47.777202812895,longitude:-122.192403046559),
-CLLocationCoordinate2D(latitude:47.7772385079546,longitude:-122.192418098025),
-CLLocationCoordinate2D(latitude:47.77727103012,longitude:-122.192441117914),
-CLLocationCoordinate2D(latitude:47.777316640474,longitude:-122.192483616171),
-CLLocationCoordinate2D(latitude:47.777330720192,longitude:-122.192532090214),
-CLLocationCoordinate2D(latitude:47.7773275472978,longitude:-122.19258499116),
-CLLocationCoordinate2D(latitude:47.7773190201447,longitude:-122.192639221237),
-CLLocationCoordinate2D(latitude:47.7772950251324,longitude:-122.192679506044),
-CLLocationCoordinate2D(latitude:47.7772718233436,longitude:-122.19271979085),
-CLLocationCoordinate2D(latitude:47.7772670640023,longitude:-122.1927715856),
-CLLocationCoordinate2D(latitude:47.7772803504967,longitude:-122.19283378304),
-CLLocationCoordinate2D(latitude:47.7772880844263,longitude:-122.192884471596),
-CLLocationCoordinate2D(latitude:47.7772646843317,longitude:-122.192925862597),
-CLLocationCoordinate2D(latitude:47.7772470351077,longitude:-122.192973009631),
-CLLocationCoordinate2D(latitude:47.7772329553898,longitude:-122.193021483675),
-CLLocationCoordinate2D(latitude:47.7772291875779,longitude:-122.193074385682),
-CLLocationCoordinate2D(latitude:47.777222048566,longitude:-122.193126180433),
-CLLocationCoordinate2D(latitude:47.7772198672012,longitude:-122.193179081378),
-CLLocationCoordinate2D(latitude:47.7772198672012,longitude:-122.193232204199),
-CLLocationCoordinate2D(latitude:47.7772321621662,longitude:-122.19328045849),
-CLLocationCoordinate2D(latitude:47.7772515961431,longitude:-122.193326498269),
-CLLocationCoordinate2D(latitude:47.7772732114848,longitude:-122.19336921734),
-CLLocationCoordinate2D(latitude:47.7772979997207,longitude:-122.193413929048),
-CLLocationCoordinate2D(latitude:47.7773174336976,longitude:-122.193459968826),
-CLLocationCoordinate2D(latitude:47.7773360744509,longitude:-122.193506230481),
-CLLocationCoordinate2D(latitude:47.7773531287572,longitude:-122.193572854822),
-CLLocationCoordinate2D(latitude:47.7773568965691,longitude:-122.193625977643),
-CLLocationCoordinate2D(latitude:47.7773576897926,longitude:-122.193679985845),
-CLLocationCoordinate2D(latitude:47.7773600694632,longitude:-122.193734215391),
-CLLocationCoordinate2D(latitude:47.7773685966164,longitude:-122.193786010142),
-],
-[
-CLLocationCoordinate2D(latitude:47.7761759850128,longitude:-122.196457202673),
-CLLocationCoordinate2D(latitude:47.7761962122133,longitude:-122.196411162895),
-CLLocationCoordinate2D(latitude:47.776219414002,longitude:-122.196371984814),
-CLLocationCoordinate2D(latitude:47.7762418225673,longitude:-122.196329486557),
-CLLocationCoordinate2D(latitude:47.7762674040267,longitude:-122.196292521927),
-CLLocationCoordinate2D(latitude:47.7762977448274,longitude:-122.196263747066),
-CLLocationCoordinate2D(latitude:47.7763233262868,longitude:-122.196227003781),
-CLLocationCoordinate2D(latitude:47.7763619959347,longitude:-122.196198228919),
-CLLocationCoordinate2D(latitude:47.7763899570647,longitude:-122.196163699085),
-CLLocationCoordinate2D(latitude:47.77641236563,longitude:-122.196122086209),
-CLLocationCoordinate2D(latitude:47.7764347741952,longitude:-122.196080694677),
-CLLocationCoordinate2D(latitude:47.7764550013956,longitude:-122.196028899926),
-CLLocationCoordinate2D(latitude:47.7764643217723,longitude:-122.195978211901),
-CLLocationCoordinate2D(latitude:47.7764696760313,longitude:-122.195925310425),
-CLLocationCoordinate2D(latitude:47.7764752285961,longitude:-122.195872408949),
-CLLocationCoordinate2D(latitude:47.7764728489254,longitude:-122.195805562732),
-CLLocationCoordinate2D(latitude:47.7764550013956,longitude:-122.195759522953),
-CLLocationCoordinate2D(latitude:47.7764419132071,longitude:-122.195709941654),
-CLLocationCoordinate2D(latitude:47.776426247042,longitude:-122.195648850409),
-CLLocationCoordinate2D(latitude:47.7764256521244,longitude:-122.195595948933),
-CLLocationCoordinate2D(latitude:47.7764411199835,longitude:-122.195548802429),
-CLLocationCoordinate2D(latitude:47.776452621725,longitude:-122.195498114404),
-CLLocationCoordinate2D(latitude:47.7764557946192,longitude:-122.195443884857),
-CLLocationCoordinate2D(latitude:47.7764557946192,longitude:-122.195389876655),
-CLLocationCoordinate2D(latitude:47.7764557946192,longitude:-122.195336975179),
-CLLocationCoordinate2D(latitude:47.7764603556546,longitude:-122.195265480715),
-CLLocationCoordinate2D(latitude:47.7764688828077,longitude:-122.195211472514),
-CLLocationCoordinate2D(latitude:47.7764837557492,longitude:-122.195164104665),
-CLLocationCoordinate2D(latitude:47.7764845489728,longitude:-122.195111203188),
-CLLocationCoordinate2D(latitude:47.776466701443,longitude:-122.195064056685),
-CLLocationCoordinate2D(latitude:47.7764518285014,longitude:-122.195016688836),
-CLLocationCoordinate2D(latitude:47.7764488539131,longitude:-122.194963787359),
-CLLocationCoordinate2D(latitude:47.7764488539131,longitude:-122.194901589389),
-CLLocationCoordinate2D(latitude:47.7764488539131,longitude:-122.194841826215),
-CLLocationCoordinate2D(latitude:47.7764488539131,longitude:-122.194787596669),
-CLLocationCoordinate2D(latitude:47.776452621725,longitude:-122.194734695193),
-CLLocationCoordinate2D(latitude:47.7764603556546,longitude:-122.19467935892),
-CLLocationCoordinate2D(latitude:47.7764688828077,longitude:-122.194625350719),
-CLLocationCoordinate2D(latitude:47.7764744353725,longitude:-122.194570014446),
-CLLocationCoordinate2D(latitude:47.7764821693021,longitude:-122.194492986356),
-CLLocationCoordinate2D(latitude:47.7764960507142,longitude:-122.194439863534),
-CLLocationCoordinate2D(latitude:47.7765045778673,longitude:-122.194385855333),
-CLLocationCoordinate2D(latitude:47.7765162779147,longitude:-122.194335167307),
-CLLocationCoordinate2D(latitude:47.7765301593267,longitude:-122.194265000914),
-CLLocationCoordinate2D(latitude:47.776540272927,longitude:-122.194211878093),
-CLLocationCoordinate2D(latitude:47.7765488000801,longitude:-122.194160083342),
-CLLocationCoordinate2D(latitude:47.7765597069039,longitude:-122.194109395317),
-CLLocationCoordinate2D(latitude:47.7765759679866,longitude:-122.194056493841),
-CLLocationCoordinate2D(latitude:47.7765959968812,longitude:-122.194010454062),
-CLLocationCoordinate2D(latitude:47.7766215783406,longitude:-122.193971275981),
-CLLocationCoordinate2D(latitude:47.7766650073298,longitude:-122.193941394394),
-CLLocationCoordinate2D(latitude:47.7766983227188,longitude:-122.193919481231),
-CLLocationCoordinate2D(latitude:47.7767316381078,longitude:-122.193897568067),
-CLLocationCoordinate2D(latitude:47.776768126391,longitude:-122.193896461341),
-CLLocationCoordinate2D(latitude:47.7768060028154,longitude:-122.193896461341),
-CLLocationCoordinate2D(latitude:47.7768470521339,longitude:-122.193894247891),
-CLLocationCoordinate2D(latitude:47.7768835404171,longitude:-122.193883844671),
-CLLocationCoordinate2D(latitude:47.7769192354768,longitude:-122.193866579754),
-CLLocationCoordinate2D(latitude:47.7769586983482,longitude:-122.193841125069),
-CLLocationCoordinate2D(latitude:47.7769896340666,longitude:-122.193813456933),
-CLLocationCoordinate2D(latitude:47.777022156232,longitude:-122.193789330318),
-CLLocationCoordinate2D(latitude:47.777055471621,longitude:-122.193765203704),
-CLLocationCoordinate2D(latitude:47.7770951327984,longitude:-122.193744397265),
-CLLocationCoordinate2D(latitude:47.7771330092228,longitude:-122.193729567145),
-CLLocationCoordinate2D(latitude:47.777169497506,longitude:-122.193724918898),
-CLLocationCoordinate2D(latitude:47.7772051925656,longitude:-122.193724918898),
-CLLocationCoordinate2D(latitude:47.7772406893194,longitude:-122.193724918898),
-CLLocationCoordinate2D(latitude:47.7772740047083,longitude:-122.193750152238),
-CLLocationCoordinate2D(latitude:47.7773073200973,longitude:-122.193772065402),
-CLLocationCoordinate2D(latitude:47.777337660898,longitude:-122.193812350208),
-],
-[
-CLLocationCoordinate2D(latitude:47.7773547152043,longitude:-122.193797520087),
-CLLocationCoordinate2D(latitude:47.7773872373697,longitude:-122.193821646701),
-CLLocationCoordinate2D(latitude:47.7774221392058,longitude:-122.193838911618),
-CLLocationCoordinate2D(latitude:47.7774848038661,longitude:-122.193882737946),
-CLLocationCoordinate2D(latitude:47.777516532808,longitude:-122.193910184737),
-CLLocationCoordinate2D(latitude:47.7775484600558,longitude:-122.193936746148),
-CLLocationCoordinate2D(latitude:47.7775994246687,longitude:-122.193981679201),
-CLLocationCoordinate2D(latitude:47.7776265925752,longitude:-122.194015102309),
-CLLocationCoordinate2D(latitude:47.7776537604817,longitude:-122.194049632143),
-CLLocationCoordinate2D(latitude:47.7776862826471,longitude:-122.194073758758),
-CLLocationCoordinate2D(latitude:47.7777249522951,longitude:-122.194063355538),
-CLLocationCoordinate2D(latitude:47.7777590609076,longitude:-122.194046090621),
-CLLocationCoordinate2D(latitude:47.7777901949318,longitude:-122.19401731576),
-CLLocationCoordinate2D(latitude:47.7778227170973,longitude:-122.193995402596),
-CLLocationCoordinate2D(latitude:47.7778536528156,longitude:-122.1940312605),
-CLLocationCoordinate2D(latitude:47.777879234275,longitude:-122.194068003785),
-CLLocationCoordinate2D(latitude:47.7779040225109,longitude:-122.194109395317),
-CLLocationCoordinate2D(latitude:47.777933570088,longitude:-122.194142818425),
-CLLocationCoordinate2D(latitude:47.7779799736656,longitude:-122.194165838314),
-CLLocationCoordinate2D(latitude:47.7780164619487,longitude:-122.194172700012),
-CLLocationCoordinate2D(latitude:47.778052751926,longitude:-122.194169379836),
-CLLocationCoordinate2D(latitude:47.7780900334328,longitude:-122.19416827311),
-CLLocationCoordinate2D(latitude:47.7781249352689,longitude:-122.194186644753),
-CLLocationCoordinate2D(latitude:47.7781521031754,longitude:-122.194221174587),
-CLLocationCoordinate2D(latitude:47.7781691574816,longitude:-122.194268321091),
-CLLocationCoordinate2D(latitude:47.7782046542354,longitude:-122.19434092228),
-CLLocationCoordinate2D(latitude:47.7782318221419,longitude:-122.194375452113),
-CLLocationCoordinate2D(latitude:47.7782659307544,longitude:-122.194394930481),
-CLLocationCoordinate2D(latitude:47.7783038071788,longitude:-122.194419278441),
-CLLocationCoordinate2D(latitude:47.7783387090149,longitude:-122.194441191605),
-CLLocationCoordinate2D(latitude:47.778373610851,longitude:-122.194459563247),
-CLLocationCoordinate2D(latitude:47.7784085126871,longitude:-122.194475721439),
-CLLocationCoordinate2D(latitude:47.7784626501942,longitude:-122.19450316823),
-CLLocationCoordinate2D(latitude:47.7784967588067,longitude:-122.194525081393),
-CLLocationCoordinate2D(latitude:47.7785324538663,longitude:-122.194545887832),
-CLLocationCoordinate2D(latitude:47.7785673557024,longitude:-122.1945653662),
-CLLocationCoordinate2D(latitude:47.7785998778679,longitude:-122.194588386089),
-CLLocationCoordinate2D(latitude:47.7786401339629,longitude:-122.194611405978),
-CLLocationCoordinate2D(latitude:47.7786704747636,longitude:-122.194637967389),
-CLLocationCoordinate2D(latitude:47.7787014104819,longitude:-122.194667848976),
-CLLocationCoordinate2D(latitude:47.7787339326474,longitude:-122.19470370688),
-CLLocationCoordinate2D(latitude:47.7787595141068,longitude:-122.194743991686),
-CLLocationCoordinate2D(latitude:47.7787805345308,longitude:-122.19479113819),
-CLLocationCoordinate2D(latitude:47.7787975888371,longitude:-122.19484071949),
-CLLocationCoordinate2D(latitude:47.7788098838021,longitude:-122.194890079445),
-CLLocationCoordinate2D(latitude:47.7788192041787,longitude:-122.194944308992),
-CLLocationCoordinate2D(latitude:47.7788261448848,longitude:-122.194999423919),
-CLLocationCoordinate2D(latitude:47.778836258485,longitude:-122.195051440015),
-CLLocationCoordinate2D(latitude:47.7788493466735,longitude:-122.195100799969),
-CLLocationCoordinate2D(latitude:47.7788834552861,longitude:-122.195115851435),
-CLLocationCoordinate2D(latitude:47.7789167706751,longitude:-122.195100799969),
-CLLocationCoordinate2D(latitude:47.778948499617,longitude:-122.195073131833),
-CLLocationCoordinate2D(latitude:47.7789834014531,longitude:-122.195051440015),
-CLLocationCoordinate2D(latitude:47.7790198897362,longitude:-122.195042143521),
-CLLocationCoordinate2D(latitude:47.7790563780194,longitude:-122.195042143521),
-CLLocationCoordinate2D(latitude:47.7791035748205,longitude:-122.195042143521),
-CLLocationCoordinate2D(latitude:47.7791446241391,longitude:-122.195051440015),
-CLLocationCoordinate2D(latitude:47.7791787327516,longitude:-122.195074459904),
-CLLocationCoordinate2D(latitude:47.7792074871052,longitude:-122.195107661667),
-CLLocationCoordinate2D(latitude:47.7792421906354,longitude:-122.195153922791),
-CLLocationCoordinate2D(latitude:47.7792762992479,longitude:-122.195167646186),
-CLLocationCoordinate2D(latitude:47.7793127875311,longitude:-122.195157242967),
-CLLocationCoordinate2D(latitude:47.7793461029201,longitude:-122.195133116352),
-CLLocationCoordinate2D(latitude:47.7793825912033,longitude:-122.195103234765),
-CLLocationCoordinate2D(latitude:47.7794151133687,longitude:-122.195082428327),
-CLLocationCoordinate2D(latitude:47.7794508084284,longitude:-122.195081321601),
-CLLocationCoordinate2D(latitude:47.7794833305938,longitude:-122.195104341491),
-CLLocationCoordinate2D(latitude:47.7795150595357,longitude:-122.195151487995),
-CLLocationCoordinate2D(latitude:47.779544408807,longitude:-122.195184911103),
-CLLocationCoordinate2D(latitude:47.7795771292783,longitude:-122.195206824267),
-CLLocationCoordinate2D(latitude:47.7796219464087,longitude:-122.195219440937),
-CLLocationCoordinate2D(latitude:47.7796576414684,longitude:-122.195219440937),
-CLLocationCoordinate2D(latitude:47.779693336528,longitude:-122.195211472514),
-CLLocationCoordinate2D(latitude:47.7797218925757,longitude:-122.195243567552),
-CLLocationCoordinate2D(latitude:47.779766908012,longitude:-122.195273670484),
-CLLocationCoordinate2D(latitude:47.7798018098481,longitude:-122.195292042126),
-CLLocationCoordinate2D(latitude:47.7798343320136,longitude:-122.195311520494),
-CLLocationCoordinate2D(latitude:47.7798985831209,longitude:-122.195357560272),
-CLLocationCoordinate2D(latitude:47.7799289239216,longitude:-122.195384121683),
-CLLocationCoordinate2D(latitude:47.7799582731928,longitude:-122.195415109995),
-CLLocationCoordinate2D(latitude:47.7799923818054,longitude:-122.195433702983),
-CLLocationCoordinate2D(latitude:47.780028076865,longitude:-122.195465798021),
-CLLocationCoordinate2D(latitude:47.780054451548,longitude:-122.195503869376),
-CLLocationCoordinate2D(latitude:47.7800715058542,longitude:-122.19555101588),
-CLLocationCoordinate2D(latitude:47.7800891550782,longitude:-122.195605245427),
-CLLocationCoordinate2D(latitude:47.7801155297611,longitude:-122.19563977526),
-CLLocationCoordinate2D(latitude:47.7801504315972,longitude:-122.19565017848),
-CLLocationCoordinate2D(latitude:47.7802117081162,longitude:-122.195638668535),
-],
-[
-CLLocationCoordinate2D(latitude:47.7802333234579,longitude:-122.19563977526),
-CLLocationCoordinate2D(latitude:47.780268225294,longitude:-122.19562583052),
-CLLocationCoordinate2D(latitude:47.7803023339065,longitude:-122.195608565603),
-CLLocationCoordinate2D(latitude:47.7803410035545,longitude:-122.195576470565),
-CLLocationCoordinate2D(latitude:47.7803713443552,longitude:-122.195548802429),
-CLLocationCoordinate2D(latitude:47.7804030732971,longitude:-122.195522241018),
-CLLocationCoordinate2D(latitude:47.7804441226156,longitude:-122.195493466157),
-CLLocationCoordinate2D(latitude:47.7804798176753,longitude:-122.195485497734),
-CLLocationCoordinate2D(latitude:47.7805155127349,longitude:-122.195485497734),
-CLLocationCoordinate2D(latitude:47.7805510094887,longitude:-122.195493466157),
-CLLocationCoordinate2D(latitude:47.7805843248776,longitude:-122.195521134293),
-CLLocationCoordinate2D(latitude:47.7806061385252,longitude:-122.195562525824),
-CLLocationCoordinate2D(latitude:47.7806325132082,longitude:-122.195599490454),
-CLLocationCoordinate2D(latitude:47.7806541285498,longitude:-122.195646636958),
-CLLocationCoordinate2D(latitude:47.780676537115,longitude:-122.195693783462),
-CLLocationCoordinate2D(latitude:47.7807013253509,longitude:-122.195731854817),
-CLLocationCoordinate2D(latitude:47.7807277000338,longitude:-122.195767491377),
-CLLocationCoordinate2D(latitude:47.7807594289757,longitude:-122.195791617991),
-CLLocationCoordinate2D(latitude:47.7808020647414,longitude:-122.195815965951),
-CLLocationCoordinate2D(latitude:47.7808377598011,longitude:-122.195817072676),
-CLLocationCoordinate2D(latitude:47.7808673073782,longitude:-122.195788297815),
-CLLocationCoordinate2D(latitude:47.7808897159434,longitude:-122.195730748092),
-CLLocationCoordinate2D(latitude:47.7809020109084,longitude:-122.195680060067),
-CLLocationCoordinate2D(latitude:47.780904390579,longitude:-122.19562583052),
-CLLocationCoordinate2D(latitude:47.7809020109084,longitude:-122.195572929044),
-CLLocationCoordinate2D(latitude:47.7808873362728,longitude:-122.19551316587),
-CLLocationCoordinate2D(latitude:47.7808849566021,longitude:-122.195460043048),
-CLLocationCoordinate2D(latitude:47.7808849566021,longitude:-122.195407141572),
-CLLocationCoordinate2D(latitude:47.7808881294963,longitude:-122.195352912025),
-CLLocationCoordinate2D(latitude:47.7808920956141,longitude:-122.195300010549),
-CLLocationCoordinate2D(latitude:47.7809014159907,longitude:-122.195248215798),
-],
-[
-CLLocationCoordinate2D(latitude:47.7838381278692,longitude:-122.198484945031),
-CLLocationCoordinate2D(latitude:47.7838056057038,longitude:-122.198463031867),
-CLLocationCoordinate2D(latitude:47.7837853785033,longitude:-122.19841920554),
-CLLocationCoordinate2D(latitude:47.783782403915,longitude:-122.198365197338),
-CLLocationCoordinate2D(latitude:47.783782403915,longitude:-122.198312074517),
-CLLocationCoordinate2D(latitude:47.7837939056564,longitude:-122.198261607837),
-CLLocationCoordinate2D(latitude:47.7838024328096,longitude:-122.198210919811),
-CLLocationCoordinate2D(latitude:47.7838024328096,longitude:-122.198139425348),
-CLLocationCoordinate2D(latitude:47.7837970785506,longitude:-122.198086523871),
-CLLocationCoordinate2D(latitude:47.7837939056564,longitude:-122.198033622395),
-CLLocationCoordinate2D(latitude:47.7837877581739,longitude:-122.197980499574),
-CLLocationCoordinate2D(latitude:47.7837861717268,longitude:-122.197927598098),
-CLLocationCoordinate2D(latitude:47.7837861717268,longitude:-122.197874696621),
-CLLocationCoordinate2D(latitude:47.7837861717268,longitude:-122.1978215738),
-CLLocationCoordinate2D(latitude:47.7837861717268,longitude:-122.197760703901),
-CLLocationCoordinate2D(latitude:47.7837707038677,longitude:-122.197713336052),
-CLLocationCoordinate2D(latitude:47.7837435359612,longitude:-122.197680134288),
-CLLocationCoordinate2D(latitude:47.7837080392074,longitude:-122.197667296273),
-CLLocationCoordinate2D(latitude:47.7836739305949,longitude:-122.197650031356),
-CLLocationCoordinate2D(latitude:47.7836414084294,longitude:-122.197625904742),
-CLLocationCoordinate2D(latitude:47.7836110676287,longitude:-122.19759712988),
-CLLocationCoordinate2D(latitude:47.7835817183575,longitude:-122.197567248293),
-CLLocationCoordinate2D(latitude:47.7835553436745,longitude:-122.197531390389),
-CLLocationCoordinate2D(latitude:47.7835273825445,longitude:-122.197484243885),
-CLLocationCoordinate2D(latitude:47.7835087417911,longitude:-122.197439310832),
-CLLocationCoordinate2D(latitude:47.783487919673,longitude:-122.197394377779),
-CLLocationCoordinate2D(latitude:47.7834676924726,longitude:-122.197349666071),
-CLLocationCoordinate2D(latitude:47.7834407228719,longitude:-122.197287468101),
-CLLocationCoordinate2D(latitude:47.783421288895,longitude:-122.197242535048),
-CLLocationCoordinate2D(latitude:47.783400268471,longitude:-122.19719870872),
-CLLocationCoordinate2D(latitude:47.7833754802352,longitude:-122.197158423914),
-CLLocationCoordinate2D(latitude:47.7833429580697,longitude:-122.197118139108),
-CLLocationCoordinate2D(latitude:47.7833112291278,longitude:-122.197088257521),
-CLLocationCoordinate2D(latitude:47.7832779137388,longitude:-122.197062802836),
-CLLocationCoordinate2D(latitude:47.7832453915734,longitude:-122.197039782947),
-CLLocationCoordinate2D(latitude:47.7832003761371,longitude:-122.197012114811),
-CLLocationCoordinate2D(latitude:47.7831662675245,longitude:-122.196998391415),
-CLLocationCoordinate2D(latitude:47.7831313656884,longitude:-122.196980019773),
-CLLocationCoordinate2D(latitude:47.7830972570759,longitude:-122.196962754856),
-CLLocationCoordinate2D(latitude:47.7830554145338,longitude:-122.196938406896),
-CLLocationCoordinate2D(latitude:47.7830300313803,longitude:-122.196905205133),
-CLLocationCoordinate2D(latitude:47.7830306262979,longitude:-122.196850975586),
-CLLocationCoordinate2D(latitude:47.7830369720863,longitude:-122.19679807411),
-CLLocationCoordinate2D(latitude:47.7830369720863,longitude:-122.196743844563),
-CLLocationCoordinate2D(latitude:47.7830183313329,longitude:-122.196700239581),
-CLLocationCoordinate2D(latitude:47.7829959227677,longitude:-122.196657519978),
-CLLocationCoordinate2D(latitude:47.7829663751906,longitude:-122.196619669968),
-CLLocationCoordinate2D(latitude:47.7829407937312,longitude:-122.196582705338),
-CLLocationCoordinate2D(latitude:47.7829082715657,longitude:-122.196559685449),
-CLLocationCoordinate2D(latitude:47.7828717832825,longitude:-122.196547068779),
-CLLocationCoordinate2D(latitude:47.7828347000817,longitude:-122.196535558835),
-CLLocationCoordinate2D(latitude:47.7827990050221,longitude:-122.196513645671),
-CLLocationCoordinate2D(latitude:47.7827688625273,longitude:-122.196480222562),
-CLLocationCoordinate2D(latitude:47.7827424878443,longitude:-122.196445692729),
-CLLocationCoordinate2D(latitude:47.7826982656315,longitude:-122.196427321086),
-CLLocationCoordinate2D(latitude:47.7826617773484,longitude:-122.196415811142),
-CLLocationCoordinate2D(latitude:47.7826278670417,longitude:-122.19639965295),
-CLLocationCoordinate2D(latitude:47.7825937584292,longitude:-122.196377739786),
-CLLocationCoordinate2D(latitude:47.7825564769224,longitude:-122.196347858199),
-CLLocationCoordinate2D(latitude:47.7825277225688,longitude:-122.196316648542),
-CLLocationCoordinate2D(latitude:47.7824983732976,longitude:-122.196285660229),
-CLLocationCoordinate2D(latitude:47.782469618944,longitude:-122.196253343846),
-CLLocationCoordinate2D(latitude:47.7824317425196,longitude:-122.196216600562),
-CLLocationCoordinate2D(latitude:47.7823984271306,longitude:-122.196198228919),
-CLLocationCoordinate2D(latitude:47.7823611456239,longitude:-122.196191145876),
-CLLocationCoordinate2D(latitude:47.7823248556466,longitude:-122.196191145876),
-CLLocationCoordinate2D(latitude:47.7822844012457,longitude:-122.196185390904),
-CLLocationCoordinate2D(latitude:47.7822496977155,longitude:-122.196167019262),
-CLLocationCoordinate2D(latitude:47.7822163823265,longitude:-122.196142892647),
-CLLocationCoordinate2D(latitude:47.782162839737,longitude:-122.196100173045),
-CLLocationCoordinate2D(latitude:47.7821311107951,longitude:-122.196077153156),
-CLLocationCoordinate2D(latitude:47.782094622512,longitude:-122.196069184733),
-CLLocationCoordinate2D(latitude:47.7820589274523,longitude:-122.196069184733),
-CLLocationCoordinate2D(latitude:47.7820101442042,longitude:-122.196069184733),
-CLLocationCoordinate2D(latitude:47.7819744491445,longitude:-122.196055461337),
-CLLocationCoordinate2D(latitude:47.7819411337555,longitude:-122.196031113377),
-CLLocationCoordinate2D(latitude:47.7818971098486,longitude:-122.19598861512),
-CLLocationCoordinate2D(latitude:47.7818699419421,longitude:-122.195952978561),
-CLLocationCoordinate2D(latitude:47.7818397994473,longitude:-122.195922875629),
-CLLocationCoordinate2D(latitude:47.7818102518702,longitude:-122.195890780591),
-CLLocationCoordinate2D(latitude:47.7817822907401,longitude:-122.195857357483),
-],
-[
-CLLocationCoordinate2D(latitude:47.7808827752374,longitude:-122.195448533104),
-CLLocationCoordinate2D(latitude:47.780876429449,longitude:-122.195500327854),
-CLLocationCoordinate2D(latitude:47.7808609615898,longitude:-122.195552122605),
-CLLocationCoordinate2D(latitude:47.7808454937306,longitude:-122.19560059718),
-CLLocationCoordinate2D(latitude:47.7808353801304,longitude:-122.19565239193),
-CLLocationCoordinate2D(latitude:47.7808276462008,longitude:-122.195705293407),
-CLLocationCoordinate2D(latitude:47.7808268529773,longitude:-122.195758416228),
-CLLocationCoordinate2D(latitude:47.7808316123186,longitude:-122.19581463788),
-CLLocationCoordinate2D(latitude:47.7808502530719,longitude:-122.195859570934),
-CLLocationCoordinate2D(latitude:47.7808780158961,longitude:-122.195891887316),
-CLLocationCoordinate2D(latitude:47.7808942769788,longitude:-122.19593903382),
-CLLocationCoordinate2D(latitude:47.7808990363201,longitude:-122.195992156642),
-CLLocationCoordinate2D(latitude:47.7809067702497,longitude:-122.196043951392),
-CLLocationCoordinate2D(latitude:47.7809230313324,longitude:-122.196092204622),
-CLLocationCoordinate2D(latitude:47.7809624942039,longitude:-122.196132489428),
-CLLocationCoordinate2D(latitude:47.78099739604,longitude:-122.196141785922),
-CLLocationCoordinate2D(latitude:47.7810338843232,longitude:-122.196141785922),
-CLLocationCoordinate2D(latitude:47.7810810811242,longitude:-122.196146434169),
-CLLocationCoordinate2D(latitude:47.7811167761839,longitude:-122.196139351126),
-CLLocationCoordinate2D(latitude:47.7811508847964,longitude:-122.196149754345),
-CLLocationCoordinate2D(latitude:47.7811842001854,longitude:-122.196169454058),
-CLLocationCoordinate2D(latitude:47.781219895245,longitude:-122.196190039151),
-CLLocationCoordinate2D(latitude:47.7812553919988,longitude:-122.196193580672),
-CLLocationCoordinate2D(latitude:47.7812902938349,longitude:-122.196205090617),
-CLLocationCoordinate2D(latitude:47.7813251956709,longitude:-122.196191145876),
-CLLocationCoordinate2D(latitude:47.7813158752943,longitude:-122.1961382444),
-CLLocationCoordinate2D(latitude:47.7813190481885,longitude:-122.196085342924),
-CLLocationCoordinate2D(latitude:47.7813515703539,longitude:-122.196064536486),
-CLLocationCoordinate2D(latitude:47.7813878603312,longitude:-122.196032441448),
-CLLocationCoordinate2D(latitude:47.7814166146848,longitude:-122.196000125065),
-CLLocationCoordinate2D(latitude:47.7814523097444,longitude:-122.195994370093),
-CLLocationCoordinate2D(latitude:47.7814794776509,longitude:-122.196018496707),
-CLLocationCoordinate2D(latitude:47.7814677776036,longitude:-122.196086449649),
-CLLocationCoordinate2D(latitude:47.7814638114859,longitude:-122.196139351126),
-CLLocationCoordinate2D(latitude:47.7814800725686,longitude:-122.196160157564),
-CLLocationCoordinate2D(latitude:47.7815104133693,longitude:-122.196118766032),
-CLLocationCoordinate2D(latitude:47.7815290541226,longitude:-122.196073832979),
-CLLocationCoordinate2D(latitude:47.7815367880522,longitude:-122.196022038229),
-CLLocationCoordinate2D(latitude:47.7815445219818,longitude:-122.195968915407),
-CLLocationCoordinate2D(latitude:47.7815693102177,longitude:-122.195918448727),
-CLLocationCoordinate2D(latitude:47.7816050052773,longitude:-122.195908045508),
-CLLocationCoordinate2D(latitude:47.781638915584,longitude:-122.195885025619),
-CLLocationCoordinate2D(latitude:47.78167381742,longitude:-122.195872408949),
-CLLocationCoordinate2D(latitude:47.7817079260326,longitude:-122.195858464208),
-],
-[
-CLLocationCoordinate2D(latitude:47.7762396412025,longitude:-122.196615021721),
-CLLocationCoordinate2D(latitude:47.7762767244033,longitude:-122.196609266749),
-CLLocationCoordinate2D(latitude:47.7763116262394,longitude:-122.196596650079),
-CLLocationCoordinate2D(latitude:47.7763201533926,longitude:-122.196566547147),
-CLLocationCoordinate2D(latitude:47.7763062719805,longitude:-122.196518293918),
-CLLocationCoordinate2D(latitude:47.7762844583329,longitude:-122.196476902386),
-CLLocationCoordinate2D(latitude:47.7762681972502,longitude:-122.196430641262),
-CLLocationCoordinate2D(latitude:47.7762473751321,longitude:-122.196382388033),
-],
-[
-CLLocationCoordinate2D(latitude:47.7764363606423,longitude:-122.195781436117),
-CLLocationCoordinate2D(latitude:47.776400070665,longitude:-122.195772139623),
-CLLocationCoordinate2D(latitude:47.7763635823818,longitude:-122.195761736404),
-CLLocationCoordinate2D(latitude:47.776319360169,longitude:-122.195749119734),
-CLLocationCoordinate2D(latitude:47.7762745430386,longitude:-122.195742258036),
-],
-[
-CLLocationCoordinate2D(latitude:47.783293381598,longitude:-122.19712389408),
-CLLocationCoordinate2D(latitude:47.7832763272917,longitude:-122.197171040584),
-CLLocationCoordinate2D(latitude:47.783260066209,longitude:-122.197219515159),
-CLLocationCoordinate2D(latitude:47.7832422186792,longitude:-122.197266661663),
-CLLocationCoordinate2D(latitude:47.7832229830082,longitude:-122.197318456413),
-],
-[
-CLLocationCoordinate2D(latitude:47.7837568224556,longitude:-122.197349666071),
-CLLocationCoordinate2D(latitude:47.78374214782,longitude:-122.197401460822),
-CLLocationCoordinate2D(latitude:47.7837328274433,longitude:-122.197454362298),
-CLLocationCoordinate2D(latitude:47.783721127396,longitude:-122.197526963487),
-CLLocationCoordinate2D(latitude:47.7837179545018,longitude:-122.197580971689),
-CLLocationCoordinate2D(latitude:47.7837165663606,longitude:-122.197633873165),
-],
-[
-CLLocationCoordinate2D(latitude:47.7836731373713,longitude:-122.197450820776),
-CLLocationCoordinate2D(latitude:47.7836731373713,longitude:-122.197453255572),
-CLLocationCoordinate2D(latitude:47.7837024866426,longitude:-122.197484243885),
-CLLocationCoordinate2D(latitude:47.7837328274433,longitude:-122.197515453542),
-CLLocationCoordinate2D(latitude:47.7837738767619,longitude:-122.197541793608),
-CLLocationCoordinate2D(latitude:47.7838056057038,longitude:-122.197566141568),
-CLLocationCoordinate2D(latitude:47.7838303939396,longitude:-122.197600671402),
-CLLocationCoordinate2D(latitude:47.7838210735629,longitude:-122.197652466152),
-CLLocationCoordinate2D(latitude:47.7838063989273,longitude:-122.197711122601),
-CLLocationCoordinate2D(latitude:47.7837925175152,longitude:-122.19775937583),
-],
-]
-
