--- PhotoPoints/View/Map/Map Feature Data/WetLands.swift
@@ -1,233 +1,0 @@
-//
-//  WetLands.swift
-//  CoordinateDrawer
-//
-//  Created by Clay Suttner on 12/22/20.
-//
-
-import Foundation
-import MapKit
-
-class WetLandGon: MKPolygon {}
-
-let wetLands: [[CLLocationCoordinate2D]] = [
-[
-CLLocationCoordinate2D(latitude:47.7743938100076,longitude:-122.191819359029),
-CLLocationCoordinate2D(latitude:47.7744076914197,longitude:-122.191867612258),
-CLLocationCoordinate2D(latitude:47.774412450761,longitude:-122.191920735079),
-CLLocationCoordinate2D(latitude:47.77441780502,longitude:-122.192004625399),
-CLLocationCoordinate2D(latitude:47.7744162185729,longitude:-122.19205774822),
-CLLocationCoordinate2D(latitude:47.7744092778668,longitude:-122.192110649165),
-CLLocationCoordinate2D(latitude:47.7743969829018,longitude:-122.192202728722),
-CLLocationCoordinate2D(latitude:47.7743860760781,longitude:-122.192254523473),
-CLLocationCoordinate2D(latitude:47.7743767557014,longitude:-122.192306318224),
-CLLocationCoordinate2D(latitude:47.7743636675128,longitude:-122.192358112975),
-CLLocationCoordinate2D(latitude:47.7743527606891,longitude:-122.192411235796),
-CLLocationCoordinate2D(latitude:47.774338879277,longitude:-122.192459489025),
-CLLocationCoordinate2D(latitude:47.7743093316998,longitude:-122.192484722365),
-CLLocationCoordinate2D(latitude:47.7742805773462,longitude:-122.192452627858),
-CLLocationCoordinate2D(latitude:47.7742629281223,longitude:-122.192405480824),
-CLLocationCoordinate2D(latitude:47.7742621348988,longitude:-122.192352358002),
-CLLocationCoordinate2D(latitude:47.7742621348988,longitude:-122.192299457057),
-CLLocationCoordinate2D(latitude:47.7742720501931,longitude:-122.192248768501),
-CLLocationCoordinate2D(latitude:47.7742906909465,longitude:-122.192189005327),
-CLLocationCoordinate2D(latitude:47.7743132978176,longitude:-122.19214739245),
-CLLocationCoordinate2D(latitude:47.7743333267121,longitude:-122.192102458866),
-CLLocationCoordinate2D(latitude:47.774338879277,longitude:-122.192026538032),
-CLLocationCoordinate2D(latitude:47.7743047706644,longitude:-122.192035834525),
-CLLocationCoordinate2D(latitude:47.7742714552754,longitude:-122.192050665177),
-CLLocationCoordinate2D(latitude:47.7742720501931,longitude:-122.191999976621),
-CLLocationCoordinate2D(latitude:47.7742877163582,longitude:-122.191951723392),
-CLLocationCoordinate2D(latitude:47.77429148417,longitude:-122.191898821385),
-CLLocationCoordinate2D(latitude:47.7743000113232,longitude:-122.191849240085),
-CLLocationCoordinate2D(latitude:47.7743295589003,longitude:-122.191819359029),
-],
-[
-CLLocationCoordinate2D(latitude:47.7739615031743,longitude:-122.196141785922),
-CLLocationCoordinate2D(latitude:47.7739266013382,longitude:-122.196148647619),
-CLLocationCoordinate2D(latitude:47.7739032012436,longitude:-122.196119872758),
-CLLocationCoordinate2D(latitude:47.7738932859492,longitude:-122.196068078007),
-CLLocationCoordinate2D(latitude:47.77387781809,longitude:-122.196017389982),
-CLLocationCoordinate2D(latitude:47.7738599705602,longitude:-122.195970243478),
-CLLocationCoordinate2D(latitude:47.7738351823244,longitude:-122.195931065397),
-CLLocationCoordinate2D(latitude:47.7738141619004,longitude:-122.19588723907),
-CLLocationCoordinate2D(latitude:47.7737877872174,longitude:-122.195833230868),
-CLLocationCoordinate2D(latitude:47.7737699396876,longitude:-122.195785863019),
-CLLocationCoordinate2D(latitude:47.7737600243932,longitude:-122.195735174994),
-CLLocationCoordinate2D(latitude:47.7737600243932,longitude:-122.195682273517),
-CLLocationCoordinate2D(latitude:47.7737933397822,longitude:-122.195673198369),
-CLLocationCoordinate2D(latitude:47.7738266551712,longitude:-122.195693783462),
-CLLocationCoordinate2D(latitude:47.7738607637838,longitude:-122.195712155104),
-CLLocationCoordinate2D(latitude:47.773895467314,longitude:-122.195728313296),
-CLLocationCoordinate2D(latitude:47.7739311623736,longitude:-122.195742258036),
-CLLocationCoordinate2D(latitude:47.7739676506568,longitude:-122.195744471487),
-CLLocationCoordinate2D(latitude:47.7740047338576,longitude:-122.195744471487),
-CLLocationCoordinate2D(latitude:47.7740505425175,longitude:-122.195721451598),
-CLLocationCoordinate2D(latitude:47.7740862375771,longitude:-122.195715696626),
-CLLocationCoordinate2D(latitude:47.7741070596953,longitude:-122.195758416228),
-CLLocationCoordinate2D(latitude:47.7741264936722,longitude:-122.195804456006),
-CLLocationCoordinate2D(latitude:47.7741520751316,longitude:-122.195841199291),
-CLLocationCoordinate2D(latitude:47.7741822176264,longitude:-122.195872408949),
-CLLocationCoordinate2D(latitude:47.7742139465683,longitude:-122.19589985574),
-CLLocationCoordinate2D(latitude:47.7742450805925,longitude:-122.19592641715),
-CLLocationCoordinate2D(latitude:47.7742891044994,longitude:-122.195952978561),
-CLLocationCoordinate2D(latitude:47.7743154791823,longitude:-122.195990828571),
-CLLocationCoordinate2D(latitude:47.7743309470415,longitude:-122.196039303146),
-CLLocationCoordinate2D(latitude:47.774345819983,longitude:-122.196104821292),
-CLLocationCoordinate2D(latitude:47.774351174242,longitude:-122.196156616043),
-CLLocationCoordinate2D(latitude:47.7743372928299,longitude:-122.196199335645),
-CLLocationCoordinate2D(latitude:47.7743023909938,longitude:-122.196214165766),
-CLLocationCoordinate2D(latitude:47.7742629281223,longitude:-122.196228110506),
-CLLocationCoordinate2D(latitude:47.7742411144748,longitude:-122.1961878257),
-CLLocationCoordinate2D(latitude:47.774230207651,longitude:-122.196137137675),
-CLLocationCoordinate2D(latitude:47.7742040312739,longitude:-122.196074939705),
-CLLocationCoordinate2D(latitude:47.7741992719326,longitude:-122.196022038229),
-CLLocationCoordinate2D(latitude:47.7741877701912,longitude:-122.195972456929),
-CLLocationCoordinate2D(latitude:47.7741651633201,longitude:-122.195932172123),
-CLLocationCoordinate2D(latitude:47.7741342276017,longitude:-122.195902290536),
-CLLocationCoordinate2D(latitude:47.7740993257657,longitude:-122.195894100767),
-CLLocationCoordinate2D(latitude:47.774063630706,longitude:-122.195890780591),
-CLLocationCoordinate2D(latitude:47.7740326949877,longitude:-122.195916013931),
-CLLocationCoordinate2D(latitude:47.7739977931516,longitude:-122.195918448727),
-CLLocationCoordinate2D(latitude:47.773962098092,longitude:-122.195920662178),
-CLLocationCoordinate2D(latitude:47.7739482166799,longitude:-122.195966701956),
-CLLocationCoordinate2D(latitude:47.7739436556445,longitude:-122.196018496707),
-CLLocationCoordinate2D(latitude:47.773928782703,longitude:-122.196064536486),
-CLLocationCoordinate2D(latitude:47.7739529760212,longitude:-122.196126734456),
-],
-[
-CLLocationCoordinate2D(latitude:47.7761093542348,longitude:-122.196118766032),
-CLLocationCoordinate2D(latitude:47.7761141135761,longitude:-122.196066971282),
-CLLocationCoordinate2D(latitude:47.7761373153649,longitude:-122.196027793201),
-CLLocationCoordinate2D(latitude:47.7761751917893,longitude:-122.195986401669),
-CLLocationCoordinate2D(latitude:47.7762108868489,longitude:-122.195980425352),
-CLLocationCoordinate2D(latitude:47.7762465819085,longitude:-122.195980425352),
-CLLocationCoordinate2D(latitude:47.7762674040267,longitude:-122.196006986763),
-CLLocationCoordinate2D(latitude:47.7762666108031,longitude:-122.196059888239),
-CLLocationCoordinate2D(latitude:47.7762566955088,longitude:-122.19611301106),
-CLLocationCoordinate2D(latitude:47.7762263547081,longitude:-122.196139351126),
-CLLocationCoordinate2D(latitude:47.7761868918366,longitude:-122.196159050839),
-CLLocationCoordinate2D(latitude:47.776151196777,longitude:-122.196161264289),
-CLLocationCoordinate2D(latitude:47.776117881388,longitude:-122.196148647619),
-],
-[
-CLLocationCoordinate2D(latitude:47.7764581742898,longitude:-122.195716803351),
-CLLocationCoordinate2D(latitude:47.7764589675134,longitude:-122.195769926173),
-CLLocationCoordinate2D(latitude:47.776459562431,longitude:-122.195822827649),
-CLLocationCoordinate2D(latitude:47.7764728489254,longitude:-122.195871080878),
-CLLocationCoordinate2D(latitude:47.7764992236084,longitude:-122.195905610712),
-CLLocationCoordinate2D(latitude:47.7765208389501,longitude:-122.195948330314),
-CLLocationCoordinate2D(latitude:47.7765386864799,longitude:-122.195995476818),
-CLLocationCoordinate2D(latitude:47.7765597069039,longitude:-122.196058781513),
-CLLocationCoordinate2D(latitude:47.7765759679866,longitude:-122.196106149362),
-CLLocationCoordinate2D(latitude:47.7765967901047,longitude:-122.196149754345),
-CLLocationCoordinate2D(latitude:47.77661919867,longitude:-122.196191145876),
-CLLocationCoordinate2D(latitude:47.7766370461998,longitude:-122.196238513725),
-CLLocationCoordinate2D(latitude:47.7766588598473,longitude:-122.196278798532),
-CLLocationCoordinate2D(latitude:47.7766929684599,longitude:-122.196268395312),
-CLLocationCoordinate2D(latitude:47.7767044702013,longitude:-122.196229217232),
-CLLocationCoordinate2D(latitude:47.7766743277065,longitude:-122.196194687398),
-CLLocationCoordinate2D(latitude:47.7766487462471,longitude:-122.196157944113),
-CLLocationCoordinate2D(latitude:47.7766263376819,longitude:-122.196115224511),
-CLLocationCoordinate2D(latitude:47.7765999629989,longitude:-122.196071398183),
-CLLocationCoordinate2D(latitude:47.7765844951397,longitude:-122.196020931503),
-CLLocationCoordinate2D(latitude:47.7765759679866,longitude:-122.195968915407),
-CLLocationCoordinate2D(latitude:47.7765674408335,longitude:-122.195916013931),
-CLLocationCoordinate2D(latitude:47.7765565340097,longitude:-122.195857357483),
-CLLocationCoordinate2D(latitude:47.7765503865272,longitude:-122.195805562732),
-CLLocationCoordinate2D(latitude:47.7765488000801,longitude:-122.195752661256),
-CLLocationCoordinate2D(latitude:47.7765488000801,longitude:-122.195680060067),
-CLLocationCoordinate2D(latitude:47.7765456271859,longitude:-122.19562583052),
-CLLocationCoordinate2D(latitude:47.7765386864799,longitude:-122.195572929044),
-CLLocationCoordinate2D(latitude:47.7765347203621,longitude:-122.195518920842),
-CLLocationCoordinate2D(latitude:47.7765216321736,longitude:-122.195471552993),
-CLLocationCoordinate2D(latitude:47.7764875235611,longitude:-122.195461149774),
-CLLocationCoordinate2D(latitude:47.7764688828077,longitude:-122.195504976101),
-CLLocationCoordinate2D(latitude:47.7764643217723,longitude:-122.19556496062),
-CLLocationCoordinate2D(latitude:47.776480582855,longitude:-122.195612107124),
-CLLocationCoordinate2D(latitude:47.7764845489728,longitude:-122.195665008601),
-],
-[
-CLLocationCoordinate2D(latitude:47.778461262053,longitude:-122.194002485639),
-CLLocationCoordinate2D(latitude:47.7784812909475,longitude:-122.193958659311),
-CLLocationCoordinate2D(latitude:47.7785031045951,longitude:-122.19391726778),
-CLLocationCoordinate2D(latitude:47.7785215470426,longitude:-122.19386768648),
-CLLocationCoordinate2D(latitude:47.7785316606428,longitude:-122.193814785004),
-CLLocationCoordinate2D(latitude:47.7785255131603,longitude:-122.193762990253),
-CLLocationCoordinate2D(latitude:47.7785038978186,longitude:-122.193720270651),
-CLLocationCoordinate2D(latitude:47.7784727637944,longitude:-122.193662720928),
-CLLocationCoordinate2D(latitude:47.778468202759,longitude:-122.193608712726),
-CLLocationCoordinate2D(latitude:47.7784666163119,longitude:-122.193555589905),
-CLLocationCoordinate2D(latitude:47.7784511484527,longitude:-122.193504901349),
-CLLocationCoordinate2D(latitude:47.7784309212523,longitude:-122.193461296897),
-CLLocationCoordinate2D(latitude:47.7784023652046,longitude:-122.193428980514),
-CLLocationCoordinate2D(latitude:47.7783789651099,longitude:-122.193389801903),
-CLLocationCoordinate2D(latitude:47.7783650836978,longitude:-122.193340220603),
-CLLocationCoordinate2D(latitude:47.7783519955093,longitude:-122.193290861179),
-CLLocationCoordinate2D(latitude:47.7783456497209,longitude:-122.193239066428),
-CLLocationCoordinate2D(latitude:47.7783464429445,longitude:-122.193165359044),
-CLLocationCoordinate2D(latitude:47.7783565565447,longitude:-122.193114670488),
-CLLocationCoordinate2D(latitude:47.7783845176747,longitude:-122.193108915516),
-CLLocationCoordinate2D(latitude:47.7783960194162,longitude:-122.19315960301),
-CLLocationCoordinate2D(latitude:47.7783991923104,longitude:-122.193226450289),
-CLLocationCoordinate2D(latitude:47.778401571981,longitude:-122.193279351234),
-CLLocationCoordinate2D(latitude:47.7784130737225,longitude:-122.19333003979),
-CLLocationCoordinate2D(latitude:47.7784293348052,longitude:-122.193378291958),
-CLLocationCoordinate2D(latitude:47.7784557094881,longitude:-122.193413929048),
-CLLocationCoordinate2D(latitude:47.7784828773946,longitude:-122.193449786952),
-CLLocationCoordinate2D(latitude:47.7785076656305,longitude:-122.193487637493),
-CLLocationCoordinate2D(latitude:47.7785378081253,longitude:-122.193532570016),
-CLLocationCoordinate2D(latitude:47.7785572421022,longitude:-122.19357971705),
-CLLocationCoordinate2D(latitude:47.7785727099614,longitude:-122.193627083838),
-CLLocationCoordinate2D(latitude:47.778580443891,longitude:-122.193678879119),
-CLLocationCoordinate2D(latitude:47.7785812371145,longitude:-122.193731780595),
-CLLocationCoordinate2D(latitude:47.7785828235616,longitude:-122.193784682072),
-CLLocationCoordinate2D(latitude:47.7785867896793,longitude:-122.193837804893),
-CLLocationCoordinate2D(latitude:47.7785766760791,longitude:-122.193897568067),
-CLLocationCoordinate2D(latitude:47.7785518878433,longitude:-122.193936746148),
-CLLocationCoordinate2D(latitude:47.7785239267132,longitude:-122.193971275981),
-],
-[
-CLLocationCoordinate2D(latitude:47.7791384766566,longitude:-122.193433629292),
-CLLocationCoordinate2D(latitude:47.7791128951972,longitude:-122.193403747174),
-CLLocationCoordinate2D(latitude:47.7791081358559,longitude:-122.193343762124),
-CLLocationCoordinate2D(latitude:47.7791222155739,longitude:-122.193296616151),
-CLLocationCoordinate2D(latitude:47.7791571174099,longitude:-122.193298829602),
-CLLocationCoordinate2D(latitude:47.7791717920456,longitude:-122.193346197451),
-CLLocationCoordinate2D(latitude:47.7791608852218,longitude:-122.193423225542),
-],
-[
-CLLocationCoordinate2D(latitude:47.7808742480842,longitude:-122.196420459388),
-CLLocationCoordinate2D(latitude:47.7808393462481,longitude:-122.196434182784),
-CLLocationCoordinate2D(latitude:47.7808092037533,longitude:-122.196419131318),
-CLLocationCoordinate2D(latitude:47.780790563,longitude:-122.19637441961),
-CLLocationCoordinate2D(latitude:47.780764188317,longitude:-122.196347858199),
-CLLocationCoordinate2D(latitude:47.7807588340581,longitude:-122.196294956723),
-CLLocationCoordinate2D(latitude:47.780757247611,longitude:-122.196240727176),
-CLLocationCoordinate2D(latitude:47.780757247611,longitude:-122.1961878257),
-CLLocationCoordinate2D(latitude:47.7807780697291,longitude:-122.196149754345),
-CLLocationCoordinate2D(latitude:47.7808105918946,longitude:-122.196167019262),
-CLLocationCoordinate2D(latitude:47.7808385530246,longitude:-122.196201549096),
-CLLocationCoordinate2D(latitude:47.7808579870015,longitude:-122.196247588874),
-CLLocationCoordinate2D(latitude:47.78087107519,longitude:-122.196294956723),
-CLLocationCoordinate2D(latitude:47.7808966566494,longitude:-122.196330593282),
-CLLocationCoordinate2D(latitude:47.7808897159434,longitude:-122.196385929555),
-],
-[
-CLLocationCoordinate2D(latitude:47.7807146118453,longitude:-122.194587279363),
-CLLocationCoordinate2D(latitude:47.7807417797518,longitude:-122.194619595746),
-CLLocationCoordinate2D(latitude:47.780769542576,longitude:-122.19465412558),
-CLLocationCoordinate2D(latitude:47.7808060308592,longitude:-122.194693303661),
-CLLocationCoordinate2D(latitude:47.7808268529773,longitude:-122.194656339031),
-CLLocationCoordinate2D(latitude:47.7808230851654,longitude:-122.19460454428),
-CLLocationCoordinate2D(latitude:47.7808183258241,longitude:-122.194551642804),
-CLLocationCoordinate2D(latitude:47.7808207054948,longitude:-122.194498741328),
-CLLocationCoordinate2D(latitude:47.7808230851654,longitude:-122.194443405056),
-CLLocationCoordinate2D(latitude:47.7808012715179,longitude:-122.194416843645),
-CLLocationCoordinate2D(latitude:47.7807687493524,longitude:-122.194441191605),
-CLLocationCoordinate2D(latitude:47.7807394000812,longitude:-122.194469966466),
-CLLocationCoordinate2D(latitude:47.7807068779157,longitude:-122.19449055156),
-CLLocationCoordinate2D(latitude:47.7806898236095,longitude:-122.194537698064),
-],
-]
-
